### 2024-02-19 #6382 @editorialbot
[PRE REVIEW]: Re-Envisioning Numerical Information Field Theory (NIFTy.re): A Library for Gaussian Processes and Variational Inference
**Submitting author:** @Edenhofer (<a href="http://orcid.org/0000-0003-3122-4894">Gordian Edenhofer</a>)
**Repository:** https://github.com/nifty-ppl/nifty
**Branch with paper.md** (empty if default branch): paper
**Version:** v8.5
