+++ paper/paper.bib
@@ -206,22 +206,38 @@
   journal   = {The Astrophysical Journal}
 }
 @article{Leike2019,
-  title     = {Charting nearby dust clouds using Gaia data only},
-  author    = {Leike, RH and En{\ss}lin, TA},
-  journal   = {Astronomy \& Astrophysics},
-  volume    = {631},
-  pages     = {A32},
-  year      = {2019},
-  publisher = {EDP Sciences}
+  author        = {{Leike}, R.~H. and {En{\ss}lin}, T.~A.},
+  title         = {{Charting nearby dust clouds using Gaia data only}},
+  journal       = {Astronomy \& Astrophysics},
+  keywords      = {dust, extinction, local insterstellar matter, methods: data analysis, solar neighborhood, Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics},
+  year          = 2019,
+  month         = nov,
+  volume        = {631},
+  eid           = {A32},
+  pages         = {A32},
+  doi           = {10.1051/0004-6361/201935093},
+  archiveprefix = {arXiv},
+  eprint        = {1901.05971},
+  primaryclass  = {astro-ph.GA},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2019A&A...631A..32L},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
 @article{Leike2020,
-  title     = {Resolving nearby dust clouds},
-  author    = {Leike,~R.~L. and Glatzle,~M. and En{\ss}lin,~T.~A.},
-  journal   = {Astronomy \& Astrophysics},
-  volume    = {639},
-  pages     = {A138},
-  year      = {2020},
-  publisher = {EDP Sciences}
+  author        = {{Leike}, R.~H. and {Glatzle}, M. and {En{\ss}lin}, T.~A.},
+  title         = {{Resolving nearby dust clouds}},
+  journal       = {Astronomy \& Astrophysics},
+  keywords      = {methods: data analysis, dust, extinction, ISM: structure, local insterstellar matter, Astrophysics - Astrophysics of Galaxies, Physics - Computational Physics},
+  year          = 2020,
+  month         = jul,
+  volume        = {639},
+  eid           = {A138},
+  pages         = {A138},
+  doi           = {10.1051/0004-6361/202038169},
+  archiveprefix = {arXiv},
+  eprint        = {2004.06732},
+  primaryclass  = {astro-ph.GA},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2020A&A...639A.138L},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
 @misc{Edenhofer2022,
   doi       = {10.48550/ARXIV.2206.10634},
