### 2024-02-19 cd16e2b
paper.md: Give Reimar a placeholder affiliation
