### 2024-02-19 #6382 @Edenhofer
Thanks for opening the pre review issue so quickly!

I see that having the paper branch's history being unrelated to the main development branch's history is undesirable for JOSS. I merged both histories so that the paper and the code now share a history.

The paper has a "statement of need" section and Reimar not having an affiliation was intentional. Reimar left academia and his contribution to this work are not related to his employer. I added a placeholder affiliation for now to make the bot happy :) I will discuss further with Reimar what we want to put there.
