### 2024-02-19 #6382 @Edenhofer
I talked with Reimar and he prefers to have no affiliation. I thus added a new affiliation with "No Affiliation" as title. I hope this is acceptable.

P.S. I also fixed the DOI warning while fiddling with the paper.
