### 2024-02-20 #6382 @dfm
Thanks! Our preferred notation for that is ["Independent researcher"](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), but otherwise all good.
