### 2024-02-20 89becb3
paper.md: Tweak Reimar's affiliation
