+++ paper/paper.md
@@ -24,7 +24,7 @@ authors:
     affiliation: "2"
   - name: Lukas I. Scheel-Platz
     orcid: 0000-0003-0809-9634
-    affiliation: "2, 4, 6"
+    affiliation: "2, 6, 7"
   - name: Matteo Guardiani
     orcid: 0000-0002-4905-6692
     affiliation: "1, 2"
@@ -50,6 +50,8 @@ affiliations:
     index: 5
   - name: Helmholtz Zentrum München, Ingolstädter Landstraße 1, 85764 Neuherberg, Germany
     index: 6
+  - name: Technical University of Munich, Ismaninger Str. 22, 81675 München, Germany
+    index: 7
 date: 21 February 2024
 bibliography: paper.bib
 
