### 2024-02-20 05d0cee
paper.md: Tweak Lukas' affiliations
