### 2024-02-20 #6382 @Edenhofer
> [...] if you have any suggestions for potential reviewers then please mention them here in this thread [...]

I think the following JOSS reviewers might be suitable reviewer for this submission

* williamjameshandley
* zhampel
* dgerosa
* Abinashbunty
* thomaspinder
* tbmiller-astro
* apizzuto
* prashjet (no collaboration but currently visiting the University of Vienna, prashjet employer)
