### 2024-02-20 327c0e1
paper.md: Harmonize affiliations
