### 2024-02-25 fe864ae
Track half-automatically generated paper.tex
