### 2024-02-25 4563052
paper.md: Avoid excessive mathcal b/c in looks bad in TeX
