### 2024-02-25 346fcff
paper.tex: Sync
