+++ paper/paper.tex
@@ -97,7 +97,6 @@
 
 % --- Default font ------------------------------------------------------------
 \renewcommand\familydefault{\sfdefault}
-\renewcommand\rmdefault{cmr}  %set back rm
 
 % --- Style -------------------------------------------------------------------
 \renewcommand{\bibfont}{\small \sffamily}
@@ -506,11 +505,11 @@ interest to the data.
 
 \texttt{NIFTy} and \texttt{NIFTy.re} build up hierarchical models for
 the posterior. The log-posterior function reads
-\(\ln\mathcal{p(\theta|d)} := \mathcal{l}(d, f(\theta)) + \ln\mathcal{p}(\theta) + \mathrm{const}\)
-with log-likelihood \(\mathcal{l}\), forward model \(f\) mapping the
-parameters of interest \(\theta\) to the data space, and log-prior
-\(\ln\mathcal{p(\theta)}\). The goal of the inference is to draw samples
-from the posterior \(\mathcal{p}(\theta|d)\).
+\(\ln{p(\theta|d)} := \ell(d, f(\theta)) + \ln{p}(\theta) + \mathrm{const}\)
+with log-likelihood \(\ell\), forward model \(f\) mapping the parameters
+of interest \(\theta\) to the data space, and log-prior
+\(\ln{p(\theta)}\). The goal of the inference is to draw samples from
+the posterior \(p(\theta|d)\).
 
 What is considered part of the likelihood versus part of the prior is
 ill-defined. Without loss of generality, \texttt{NIFTy} and
@@ -533,7 +532,7 @@ harmonic domain being easily accessible. For example, for pixels spaced
 on a regular Cartesian grid, the natural choice to represent a
 stationary kernel is the Fourier domain. In the generative picture, a
 realization \(s\) drawn from a GP then reads
-\(s = \mathcal{FT} \circ \sqrt{P} \circ \xi\) with \(\mathcal{FT}\) the
+\(s = \mathrm{FT} \circ \sqrt{P} \circ \xi\) with \(\mathrm{FT}\) the
 (fast) Fourier transform, \(\sqrt{P}\) the square-root of the
 power-spectrum in harmonic space, and \(\xi\) standard Gaussian random
 variables. In the implementation in \texttt{NIFTy.re} and
