### 2024-02-25 de3a3cb
paper.tex: Manually tweak mathbb text
