### 2024-02-25 a778694
paper.tex: Remove left margin
