+++ paper/paper.tex
@@ -92,8 +92,8 @@
 \newcommand{\CSLIndent}[1]{\hspace{\cslhangindent}#1}
 
 % --- Page layout -------------------------------------------------------------
-\usepackage[top=3.5cm, bottom=3cm, right=1.5cm, left=1.0cm,
-            headheight=2.2cm, reversemp, includemp, marginparwidth=4.5cm]{geometry}
+\usepackage[top=3.5cm, bottom=3cm, right=2.54cm, left=2.54cm,
+            headheight=2.2cm]{geometry}
 
 % --- Default font ------------------------------------------------------------
 \renewcommand\familydefault{\sfdefault}
@@ -350,54 +350,6 @@ Munich, Ismaninger Str. 22, 81675 München, Germany}
 
 \begin{document}
 \maketitle
-
-\iffalse
-\marginpar{
-
-  \begin{flushleft}
-  %\hrule
-  \sffamily\small
-
-  {\bfseries DOI:} \href{https://doi.org/DOI unavailable}{\color{linky}{DOI unavailable}}
-
-  \vspace{2mm}
-
-  {\bfseries Software}
-  \begin{itemize}
-    \setlength\itemsep{0em}
-    \item \href{https://github.com/openjournals/joss-reviews/issues/6382}{\color{linky}{Review}} \ExternalLink
-    \item \href{https://github.com/nifty-ppl/nifty}{\color{linky}{Repository}} \ExternalLink
-    \item \href{DOI unavailable}{\color{linky}{Archive}} \ExternalLink
-  \end{itemize}
-
-  \vspace{2mm}
-
-  \par\noindent\hrulefill\par
-
-  \vspace{2mm}
-
-  {\bfseries Editor:} \href{https://github.com/openjournals/joss-reviews/issues/6382}{Pending
-Editor} \ExternalLink \\
-  \vspace{1mm}
-    {\bfseries Reviewers:}
-  \begin{itemize}
-  \setlength\itemsep{0em}
-    \item \href{https://github.com/Pending Reviewers}{@Pending
-Reviewers}
-    \end{itemize}
-    \vspace{2mm}
-
-  {\bfseries Submitted:} 19 February 2024\\
-  {\bfseries Published:} N/A
-
-  \vspace{2mm}
-  {\bfseries License}\\
-  Authors of papers retain copyright and release the work under a Creative Commons Attribution 4.0 International License (\href{http://creativecommons.org/licenses/by/4.0/}{\color{linky}{CC BY 4.0}}).
-
-
-  \end{flushleft}
-}
-\fi
 
 \vspace{1em}
 \hypertarget{summary}{%
