### 2024-02-27 4bc0ee7
Update NIFTy.re citation
