+++ src/re/refine/chart.py
@@ -6,8 +6,8 @@ from functools import partial
 from math import log2, sqrt
 from typing import Callable, Iterable, Literal, Optional, Tuple, Union
 
-from jax import numpy as jnp
 import numpy as np
+from jax import numpy as jnp
 
 from .util import (
     coarse2fine_distances,
@@ -22,7 +22,7 @@ MAX_SIZE0 = 1024
 NEST = True
 
 
-class CoordinateChart():
+class CoordinateChart:
     def __init__(
         self,
         min_shape: Optional[Iterable[int]] = None,
@@ -32,12 +32,22 @@ class CoordinateChart():
         _coarse_size: int = 5,
         _fine_size: int = 4,
         _fine_strategy: Literal["jump", "extend"] = "extend",
-        rg2cart: Optional[Callable[[
-            Iterable,
-        ], Iterable]] = None,
-        cart2rg: Optional[Callable[[
-            Iterable,
-        ], Iterable]] = None,
+        rg2cart: Optional[
+            Callable[
+                [
+                    Iterable,
+                ],
+                Iterable,
+            ]
+        ] = None,
+        cart2rg: Optional[
+            Callable[
+                [
+                    Iterable,
+                ],
+                Iterable,
+            ]
+        ] = None,
         regular_axes: Optional[Union[Iterable[int], Tuple]] = None,
         irregular_axes: Optional[Union[Iterable[int], Tuple]] = None,
         distances: Optional[Union[Iterable[float], float]] = None,
@@ -107,7 +117,7 @@ class CoordinateChart():
                     ceil_sizes=True,
                     _coarse_size=_coarse_size,
                     _fine_size=_fine_size,
-                    _fine_strategy=_fine_strategy
+                    _fine_strategy=_fine_strategy,
                 )
                 if np.prod(shape0, dtype=int) <= MAX_SIZE0:
                     break
@@ -125,17 +135,17 @@ class CoordinateChart():
                 ceil_sizes=True,
                 _coarse_size=_coarse_size,
                 _fine_size=_fine_size,
-                _fine_strategy=_fine_strategy
+                _fine_strategy=_fine_strategy,
             )
         elif shape0 is None:
             raise ValueError("either `shape0` or `min_shape` must be specified")
-        self._shape0 = (shape0, ) if isinstance(shape0, int) else tuple(shape0)
+        self._shape0 = (shape0,) if isinstance(shape0, int) else tuple(shape0)
         self._shape = coarse2fine_shape(
             shape0,
             depth,
             _coarse_size=_coarse_size,
             _fine_size=_fine_size,
-            _fine_strategy=_fine_strategy
+            _fine_strategy=_fine_strategy,
         )
 
         if _fine_strategy not in ("jump", "extend"):
@@ -147,7 +157,7 @@ class CoordinateChart():
             self.shape0,
             _coarse_size=_coarse_size,
             _fine_size=_fine_size,
-            _fine_strategy=_fine_strategy
+            _fine_strategy=_fine_strategy,
         )
 
         self._coarse_size = int(_coarse_size)
@@ -160,41 +170,37 @@ class CoordinateChart():
 
         if rg2cart is None and cart2rg is None:
             if distances0 is None and distances is None:
-                distances = jnp.ones((self.ndim, ))
+                distances = jnp.ones((self.ndim,))
                 distances0 = fine2coarse_distances(
                     distances,
                     depth,
[…]
