+++ test/test_re/test_likelihood.py
@@ -29,17 +29,14 @@ def test_partial_insert_and_remove():
     # are length one tuples
     _id_part = jpartial(
         _identity,
-        insert_axes=(jft.Vector({
-            "a": (True, False),
-            "b": False
-        }), ),
-        flat_fill=(("THIS IS input['a'][0]", ), )
+        insert_axes=(jft.Vector({"a": (True, False), "b": False}),),
+        flat_fill=(("THIS IS input['a'][0]",),),
     )
     out = _id_part(("THIS IS input['a'][1]", "THIS IS input['b']"))
     assert out == jft.Vector(
         {
             "a": ("THIS IS input['a'][0]", "THIS IS input['a'][1]"),
-            "b": "THIS IS input['b']"
+            "b": "THIS IS input['b']",
         }
     )
 
@@ -57,12 +54,7 @@ def test_likelihood_partial(seed, forward):
     key = random.PRNGKey(seed)
     forward = partial(jax.tree_map, forward)
 
-    domain = jft.Vector(
-        {
-            "a": jft.ShapeWithDtype(128),
-            "b": jft.ShapeWithDtype(64)
-        }
-    )
+    domain = jft.Vector({"a": jft.ShapeWithDtype(128), "b": jft.ShapeWithDtype(64)})
     key, sk_d, sk_p = random.split(key, 3)
     primals = jft.random_like(sk_p, domain)
     data = forward(jft.random_like(sk_d, domain))
@@ -70,7 +62,7 @@ def test_likelihood_partial(seed, forward):
     gaussian = jft.Gaussian(data).amend(forward)
 
     gaussian_part, primals_liquid = gaussian.freeze(
-        primals=primals, point_estimates=("b", )
+        primals=primals, point_estimates=("b",)
     )
     assert primals_liquid.tree[0].shape == domain["a"].shape
     aallclose(gaussian_part(primals_liquid), gaussian(primals))
@@ -95,9 +87,7 @@ def test_likelihood_partial(seed, forward):
 def test_likelihood_domain():
     lh = jft.Poissonian(jnp.zeros((3, 4), dtype=int))
     assert lh.domain is not None
-    lh_wfwd = lh.amend(
-        jft.Model(lambda x: x, domain=lh.domain, white_init=True)
-    )
+    lh_wfwd = lh.amend(jft.Model(lambda x: x, domain=lh.domain, white_init=True))
     assert lh_wfwd.domain is not None
 
 
@@ -112,8 +102,8 @@ def test_nonvariable_likelihood_add(seed, likelihood, forward_a, forward_b):
     key = random.PRNGKey(seed)
     N_TRIES = 100
     shp_a, shp_b = (3, 5), (12, 5)
-    swd_a = jft.Vector((jft.ShapeWithDtype(shp_a), ) * 2)
-    swd_b = jft.Vector((jft.ShapeWithDtype(shp_b), ) * 3)
+    swd_a = jft.Vector((jft.ShapeWithDtype(shp_a),) * 2)
+    swd_b = jft.Vector((jft.ShapeWithDtype(shp_b),) * 3)
     key_a, key_b = "a", "b"
 
     def fwd_a(x):
@@ -149,52 +139,37 @@ def test_nonvariable_likelihood_add(seed, likelihood, forward_a, forward_b):
     assert_allclose(jax.vmap(lh_orig)(p), jax.vmap(lh_ab)(p), equal_nan=False)
     rsm_orig = jax.vmap(lh_orig.right_sqrt_metric)(p, t)
[…]
