+++ test/test_re/test_forest_math.py
@@ -7,6 +7,6 @@ def test_map_forest_axes_validation():
     f = lambda x: x
     jft.map_forest(f, in_axes=1)
     with pytest.raises(ValueError):
-        jft.map_forest(f, in_axes=(None, ))
+        jft.map_forest(f, in_axes=(None,))
     with pytest.raises(TypeError):
-        jft.map_forest(f, in_axes=(1., ))
+        jft.map_forest(f, in_axes=(1.0,))
