### 2024-02-27 22ffc51
minimal_example.py: Move up imports
