### 2024-02-27 706eb61
minimal_example.py: Plot prior samples
