### 2024-02-27 f49e25e
Script a simple nonlinear regression demo using NIFTy.re
