### 2024-02-27 8c35a81
Script a simple nonlinear regression demo using NIFTy.re
