### 2024-02-27 1b34aec
paper: Fix citations
Also update paper.tex.
