+++ paper/paper.bib
@@ -269,7 +269,7 @@
 
 @article{Guardiani2022,
   doi       = {10.1371/journal.pone.0275011},
-  author    = {Guardiani, Matteo AND Frank, Philipp AND Kostić, Andrija AND Edenhofer, Gordian AND Roth, Jakob AND Uhlmann, Berit AND Enßlin, Torsten},
+  author    = {Guardiani, Matteo and Frank, Philipp and Kostić, Andrija and Edenhofer, Gordian and Roth, Jakob and Uhlmann, Berit and Enßlin, Torsten},
   journal   = {PLOS ONE},
   publisher = {Public Library of Science},
   title     = {Causal, Bayesian, & non-parametric modeling of the SARS-CoV-2 viral load distribution vs. patient’s age},
