### 2024-02-27 78ef421
z_nonlinear_regression.py: Tweak optimization parameters
