### 2024-02-27 5544db9
paper.bib: Remove duplicate citation
Update paper.tex accordingly.
