+++ paper/paper.bib
@@ -589,13 +589,6 @@
   year    = {2023}
 }
 
-@article{Blondel2021,
-  title   = {Efficient and Modular Implicit Differentiation},
-  author  = {Blondel, Mathieu and Berthet, Quentin and Cuturi, Marco and Frostig, Roy and Hoyer, Stephan and Llinares-L{\'o}pez, Felipe and Pedregosa, Fabian and Vert, Jean-Philippe},
-  journal = {arXiv preprint arXiv:2105.15183},
-  year    = {2021}
-}
-
 @article{Steiniger2017,
   author  = {{Steininger}, Theo and {Dixit}, Jait and {Frank}, Philipp and {Greiner}, Maksim and {Hutschenreuter}, Sebastian and {Knollm{\"u}ller}, Jakob and {Leike}, Reimar H. and {Porqueres}, Natalia and {Pumpe}, Daniel and {Reinecke}, Martin and {{\v{S}}raml}, Matev{\v{z}} and {Varady}, Csongor and {En{\ss}lin}, Torsten A.},
   title   = {{NIFTy 3 - Numerical Information Field Theory: A Python Framework for Multicomponent Signal Inference on HPC Clusters}},
