+++ paper/paper.tex
@@ -442,7 +442,7 @@ VI techniques (Frank et al., 2021). With \texttt{NIFTy.re}, the GP
 models and the VI machinery are now fully accessible in the JAX
 ecosystem and \texttt{NIFTy.re} components interact seamlessly with
 other JAX packages such as BlackJAX and JAXopt/Optax (Blondel et al.,
-2021; DeepMind et al., 2020).
+2022; DeepMind et al., 2020).
 
 \hypertarget{core-components}{%
 \section{Core Components}\label{core-components}}
@@ -614,7 +614,7 @@ differentiable by JAX. In addition to stand-alone sampling functions,
 \texttt{NIFTy.re} also provides tools to configure and execute the
 alternating Kullback--Leibler divergence optimization and sample
 adaption at a lower abstraction level. These tools are provided in a
-JAXopt/Optax-style optimizer class (Blondel et al., 2021; DeepMind et
+JAXopt/Optax-style optimizer class (Blondel et al., 2022; DeepMind et
 al., 2020).
 
 A typical minimization with \texttt{NIFTy.re} is shown in the following.
@@ -823,15 +823,10 @@ of Machine Learning Research}, \emph{20}, 28:1--28:6.
 
 \leavevmode\hypertarget{ref-Blondel2021}{}%
 Blondel, M., Berthet, Q., Cuturi, M., Frostig, R., Hoyer, S.,
-Llinares-López, F., Pedregosa, F., \& Vert, J.-P. (2021). Efficient and
-modular implicit differentiation. \emph{arXiv Preprint
-arXiv:2105.15183}.
-
-\leavevmode\hypertarget{ref-Blondel2021}{}%
-Blondel, M., Berthet, Q., Cuturi, M., Frostig, R., Hoyer, S.,
-Llinares-López, F., Pedregosa, F., \& Vert, J.-P. (2021). Efficient and
-modular implicit differentiation. \emph{arXiv Preprint
-arXiv:2105.15183}.
+Llinares-Lopez, F., Pedregosa, F., \& Vert, J.-P. (2022).
+\emph{Efficient and modular implicit differentiation}. \emph{35},
+5230--5242.
+\url{https://proceedings.neurips.cc/paper_files/paper/2022/file/228b9279ecf9bbafe582406850c57115-Paper-Conference.pdf}
 
 \leavevmode\hypertarget{ref-Jax2018}{}%
 Bradbury, J., Frostig, R., Hawkins, P., Johnson, M. J., Leary, C.,
