### 2024-02-28 fa8574e
Add tomography demonstration from 3D n_H repository
