### 2024-02-28 aa7e9ff
Remove outdated NIFTy.re README
