### 2024-02-28 b77f29f
Move LOS implementation from demo to NIFTy.re
