+++ src/re/__init__.py
@@ -7,9 +7,10 @@ from .custom_map import lmap, smap
 from .evi import (
     Samples, draw_linear_residual, draw_residual, nonlinearly_update_residual
 )
+from .extra import SamplingCartesianGridLOS
 from .gauss_markov import (
-    WienerProcess, IntegratedWienerProcess, OrnsteinUhlenbeckProcess,
-    GaussMarkovProcess
+    GaussMarkovProcess, IntegratedWienerProcess, OrnsteinUhlenbeckProcess,
+    WienerProcess
 )
 from .hmc import generate_hmc_acc_rej, generate_nuts_tree
 from .hmc_oo import HMCChain, NUTSChain
