### 2024-02-28 ffa40fd
Move LOS implementation from demo to NIFTy.re
