+++ src/re/extra/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+
+from .sampling_los import SamplingCartesianGridLOS
