+++ demos/1_tomography.py
@@ -1,14 +1,18 @@
-# %%
-import dataclasses
-from functools import partial
+#!/usr/bin/env python3
 
+# Copyright(C) 2013-2021 Max-Planck-Society
+# SPDX-License-Identifier: GPL-2.0+ OR BSD-2-Clause
+
+# %% [markdown]
+# # 3D Tomography Example with known start- and end-points
+
+# %%
 import jax
 import matplotlib.pyplot as plt
 import numpy as np
 import plotly.graph_objects as go
 from jax import numpy as jnp
 from jax import random
-from jax.scipy.ndimage import map_coordinates
 
 import nifty8.re as jft
 
@@ -28,86 +32,6 @@ cfm.add_fluctuations(  # Axis over which the kernle is defined
     non_parametric_kind="power",
 )
 correlated_field = cfm.finalize()  # forward model for a GP prior
-
-
-# %%
-def _los(x, /, start, end, *, distances, shape, n_sampling_points, order=1):
-    from jax.scipy.ndimage import map_coordinates
-
-    l2i = ((shape - 1) / shape) / distances
-    start_iloc = start * l2i
-    end_iloc = end * l2i
-    ddi = (end_iloc - start_iloc) / n_sampling_points
-    adi = jnp.arange(0, n_sampling_points) + 0.5
-    dist = jnp.linalg.norm(end - start)
-    pp = start_iloc[:, jnp.newaxis] + ddi[:, jnp.newaxis] * adi[jnp.newaxis]
-    return map_coordinates(x, pp, order=order, cval=jnp.nan).sum() * (
-        dist / n_sampling_points
-    )
-
-
-class LOS(jft.Model):
-    start: jax.Array = dataclasses.field(metadata=dict(static=False))
-    end: jax.Array = dataclasses.field(metadata=dict(static=False))
-    distances: jax.Array = dataclasses.field(metadata=dict(static=False))
-
-    def __init__(
-        self,
-        start,
-        end,
-        distances,
-        shape,
-        n_sampling_points=500,
-        interpolation_order=1,
-    ):
-        # We assume that `start` and `end` are of shape (n_points, n_dimensions)
-        self.start = jnp.array(start)
-        self.end = jnp.array(end)
-        self.distances = jnp.array(distances)
-        self.shape = jnp.array(shape)
-        self._los = partial(
-            _los,
-            n_sampling_points=n_sampling_points,
-            order=interpolation_order,
-            distances=self.distances,
-            shape=self.shape,
-        )
-
-    def __call__(self, x):
-        in_axes = (None, 0, 0)
-        if self.start.ndim < self.end.ndim:
-            in_axes = (None, None, 0)
-        elif self.start.ndim > self.end.ndim:
-            in_axes = (None, 0, None)
-        return jax.vmap(self._los, in_axes=in_axes)(x, self.start, self.end)
-
-
-test_key = random.PRNGKey(42)
-for test_shape in ((10,), (25,), (12, 12), (6, 6, 6)):
-    n_test_points = 1_000
-    test_los = LOS(
-        np.zeros((len(test_shape),)),
-        np.ones((len(test_shape),))[np.newaxis],
-        distances=tuple(1.0 / s for s in test_shape),
-        shape=test_shape,
-        n_sampling_points=n_test_points,
-    )
-    for test_x in (jnp.ones(test_shape), random.normal(test_key, test_shape)):
-        desired = map_coordinates(
-            test_x,
-            [
-                np.linspace(0, s - 1, num=n_test_points, endpoint=False)
-                for s in test_x.shape
-            ],
-            order=1,
[…]
