### 2024-02-28 db7a7ae
demos/1_tomography.py: Remove plotly 3D visualization
Avoid another dependency by not showing the reconstruction in 3D.
