### 2024-02-28 1c54330
demos/1_tomography.py: Tweak data visualization in projection plots
