### 2024-02-28 57a53cd
demos/{0,1}*.py: Prettify
