### 2024-03-01 c54e32e
re.msc: fix hp ordering revert to knn seq 1
