### 2024-03-01 d00da6a
re.msc.gps: jit kernelfuncs
