### 2024-03-05 c32db46
optimize_kl: Improve style
