### 2024-03-05 a1721ab
re.gauss_markov: use cumsum
