### 2024-03-05 a4458fb
optimize_kl: Make `nit` a counter instead of an iteration index
