### 2024-03-05 11de72e
re.gauss_markov: test cumsum vs fori
