### 2024-03-05 545faa4
gauss_markov: Format
