### 2024-03-05 1b5b9b3
optimize_kl: Fix off by one index in n_total_iterations
