### 2024-03-06 fd94794
minimal_benchmark.py: Adapt to new NIFTy version
