### 2024-03-07 9a6ca3c
First multi_grid skeleton for ICR and MSC
Co-authored-by: Gordian Edenhofer <gordian.edenhofer@gmail.com>
