+++ src/re/multi_grid/indexing.py
@@ -1,0 +1,293 @@
+#!/usr/bin/env python3
+
+import operator
+from dataclasses import dataclass
+from functools import reduce
+from typing import Optional, Union
+
+import numpy as np
+
+_kind_of_int = Union[tuple[int], list[int], int]
+
+
+@dataclass(kw_only=True)
+class GridAxisAtLevel:
+    size: int
+    refinement_size: int
+    parent_refinement_size: int = None
+
+    def _parse_index(self, index):
+        if np.any(np.abs(index) >= self.size):
+            nm = self.__class__.__name__
+            ve = f"index {index} is out of bounds for {nm} with size {self.size}"
+            raise IndexError(ve)
+        return index % self.size
+
+    def children(self, index) -> np.ndarray:
+        index = self._parse_index(index)
+        dtp = np.result_type(index)
+        return np.add.outer(
+            self.refinement_size * index, np.arange(self.refinement_size, dtype=dtp)
+        )
+
+    def neighborhood(self, index, window_size: int):
+        index = self._parse_index(index)
+        dtp = np.result_type(index)
+        return (
+            np.add.outer(index, np.arange(window_size, dtype=dtp) - window_size // 2)
+            % self.size
+        )
+
+    def parent(self, index):
+        if self.parent_refinement_size is None:
+            raise IndexError("you are alone in this world")
+        return index // self.parent_refinement_size
+
+
+class GridAxis:
+    """Dense (single) axis with periodic boundary conditions.
+
+    Open boundary conditions can be emulated by leaving out indices."""
+
+    def __init__(self, *, size0: int, refinement_size: _kind_of_int):
+        self.size0 = size0
+        refinement_size = (
+            (refinement_size,) if isinstance(refinement_size, int) else refinement_size
+        )
+        self.refinement_size = tuple(refinement_size)
+        self.depth = len(self.refinement_size)
+
+    def _parse_level(self, level):
+        if np.abs(level) >= self.depth:
+            raise IndexError(f"{self.__class__.__name__} does not have level {level}")
+        return level % self.depth
+
+    def amend(self, refinement_size: _kind_of_int):
+        refinement_size = (
+            (refinement_size,) if isinstance(refinement_size, int) else refinement_size
+        )
+        return GridAxis(
+            size0=self.size0,
+            refinement_size=self.refinement_size + tuple(refinement_size),
+        )
+
+    def at(self, level: int) -> GridAxisAtLevel:
+        level = self._parse_level(level)
+        size = self.size0 * reduce(operator.mul, self.refinement_size[:level], 1)
+        rs = self.refinement_size[level]
+        rs_p = self.refinement_size[level - 1] if level >= 1 else None
+        return GridAxisAtLevel(
+            size=size, refinement_size=rs, parent_refinement_size=rs_p
+        )
+
+
+def _fill_bad_healpix_neighbors(nside, neighbors, nest: bool = True):
+    import warnings
+
+    from healpy.pixelfunc import get_all_neighbours
+
+    idx_w_invalid, nbr_w_invalid = np.nonzero(neighbors == -1)
+    if idx_w_invalid.size == 0:
+        return neighbors
+
+    # Account for unknown neighbors, encoded by -1
+    uniq_idx_w_invalid = np.unique(idx_w_invalid)
+    nbr_invalid = neighbors[:, 1:][uniq_idx_w_invalid]
+    with warnings.catch_warnings():
+        wmsg = "invalid value encountered in _get_neigbors"
+        warnings.filterwarnings("ignore", message=wmsg)
[…]
