### 2024-03-07 3469ef0
SamplingCartesianGridLOS: Add doc-string
