### 2024-03-08 852d509
re.multigrid: Flesh out OGridAtLevel
