### 2024-03-08 bc9e452
demos/0_intro.py: Embrace jft.Model.init for initialization
