### 2024-03-08 b8de5b9
SamplingCartesianGridLOS: Tweak doc-string
