### 2024-03-08 9936785
Prepare for jax.tree_map -> jax.tree_util.tree_map deprecation
