+++ test/test_re/test_likelihood_impl.py
@@ -9,6 +9,7 @@ from functools import partial, reduce
 import jax
 import jax.numpy as jnp
 from jax import random
+from jax.tree_util import tree_map
 from numpy.testing import assert_allclose
 
 import nifty8.re as jft
@@ -75,7 +76,7 @@ def random_draw(key, shape, dtype, method):
             return reduce(lambda a, b: a * b, (isinstance(ii, int) for ii in x))
         return False
 
-    swd = jax.tree_map(
+    swd = tree_map(
         lambda x: jft.ShapeWithDtype(x, dtype), shape, is_leaf=_isleaf
     )
     return jft.random_like(key, jft.Vector(swd), method)
@@ -134,7 +135,7 @@ lh_init_approx = (
             "dof": partial(random_draw, dtype=float, method=random.exponential),
         }, lambda key, shape: (
             random_draw(key, shape, float, random.normal),
-            jax.tree_map(
+            tree_map(
                 jnp.exp, 3. + 1e-1 *
                 random_draw(key, shape, float, random.normal)
             )
@@ -154,7 +155,7 @@ def test_gaussian_vs_vcgaussian_consistency(seed, shape):
     m2 = random_draw(sk.pop(), shape, float, random.normal)
     t = random_draw(sk.pop(), shape, float, random.normal)
     inv_std = random_draw(sk.pop(), shape, float, random.normal)
-    inv_std = 1. / jax.tree_map(jnp.exp, 1. + inv_std)
+    inv_std = 1. / tree_map(jnp.exp, 1. + inv_std)
 
     gauss = jft.Gaussian(d, noise_std_inv=lambda x: inv_std * x)
     vcgauss = jft.VariableCovarianceGaussian(d)
@@ -162,13 +163,13 @@ def test_gaussian_vs_vcgaussian_consistency(seed, shape):
     diff_g = gauss(m2) - gauss(m1)
     diff_vcg = vcgauss((m2, inv_std)) - vcgauss((m1, inv_std))
 
-    jax.tree_map(
+    tree_map(
         partial(assert_allclose, rtol=rtol, atol=atol), diff_g, diff_vcg
     )
 
     met_g = gauss.metric(m1, t)
     met_vcg = vcgauss.metric((m1, inv_std), (t, d / 2))[0]
-    jax.tree_map(partial(assert_allclose, rtol=rtol, atol=atol), met_g, met_vcg)
+    tree_map(partial(assert_allclose, rtol=rtol, atol=atol), met_g, met_vcg)
 
 
 def test_studt_vs_vcstudt_consistency(seed, shape):
@@ -183,20 +184,20 @@ def test_studt_vs_vcstudt_consistency(seed, shape):
     m2 = random_draw(sk.pop(), shape, float, random.normal)
     t = random_draw(sk.pop(), shape, float, random.normal)
     inv_std = random_draw(sk.pop(), shape, float, random.normal)
-    inv_std = 1. / jax.tree_map(jnp.exp, 1. + inv_std)
+    inv_std = 1. / tree_map(jnp.exp, 1. + inv_std)
 
     studt = jft.StudentT(d, dof, noise_std_inv=lambda x: inv_std * x)
     vcstudt = jft.VariableCovarianceStudentT(d, dof)
 
     diff_t = studt(m2) - studt(m1)
     diff_vct = vcstudt((m2, 1. / inv_std)) - vcstudt((m1, 1. / inv_std))
-    jax.tree_map(
+    tree_map(
         partial(assert_allclose, rtol=rtol, atol=atol), diff_t, diff_vct
     )
 
     met_t = studt.metric(m1, t)
     met_vct = vcstudt.metric((m1, 1. / inv_std), (t, d / 2))[0]
[…]
