### 2024-03-09 4cbca09
.gitlab-ci.yml: Fix nonlinear pipeline
