### 2024-03-12 7f54712
Move dependency specification fully into pyproject.toml
Create a single source of truth for specifying dependencies
