### 2024-03-12 f8da7fc
README.md: Remove specification of test dependencies
This is already part of the pyproject.toml.
