### 2024-03-12 3a7c0f3
pyproject.toml: Add "all" extra dependency
