+++ pyproject.toml
@@ -63,6 +63,7 @@ test = [
     "scipy",
     "ducc0",
 ]
+all = ["nifty8[re,native,parallel,doc,util,test]"]
 
 [build-system]
 requires = [
