### 2024-03-12 8689dee
Prune custom dockerfile and instead cache pip
