### 2024-03-12 848dfe1
Re-introduce docker for LaTeX
