### 2024-03-13 6a1b560
.gitlab-ci.yaml: Re-use environment for pages
