### 2024-03-13 f78abdc
indexing.py: Revise indexing logic
Co-authored-by: Gordian Edenhofer <gordian.edenhofer@gmail.com>
