+++ src/re/multi_grid/indexing.py
@@ -3,7 +3,7 @@
 import operator
 from dataclasses import dataclass
 from functools import reduce
-from typing import Optional, Union
+from typing import Any, Callable, Iterable, Optional, Union
 
 import numpy as np
 
@@ -11,45 +11,82 @@ _kind_of_int = Union[tuple[int], list[int], int]
 
 
 @dataclass(kw_only=True)
-class GridAxisAtLevel:
-    size: int
-    refinement_size: int
-    parent_refinement_size: int = None
+class GridAtLevel:
+    shape: tuple[int]
+    refinement_shape: tuple[int]
+    parent_refinement_shape: Optional[tuple[int]] = None
+
+    def __init__(self, shape, refinement_shape, parent_refinement_shape):
+        self.shape = np.asarray(shape)
+        self.refinement_shape = np.asarray(refinement_shape)
+        self.parent_refinement_shape = np.asarray(parent_refinement_shape)
 
     def _parse_index(self, index):
-        if np.any(np.abs(index) >= self.size):
+        index = np.asarray(index)
+        if np.any(np.any(np.abs(idx) >= s) for idx, s in zip(index, self.shape)):
             nm = self.__class__.__name__
-            ve = f"index {index} is out of bounds for {nm} with size {self.size}"
+            ve = f"index {index} is out of bounds for {nm} with shape {self.shape}"
             raise IndexError(ve)
-        return index % self.size
+        return index % self.shape[(slice(None),) + (np.newaxis,) * (index.ndim - 1)]
+
+    @property
+    def size(self):
+        return len(self.shape)
 
     def children(self, index) -> np.ndarray:
         index = self._parse_index(index)
         dtp = np.result_type(index)
-        return np.add.outer(
-            self.refinement_size * index, np.arange(self.refinement_size, dtype=dtp)
+        rs = self.refinement_shape[(slice(None),) + (None,) * (index.ndim - 1)]
+        rs_bc = (slice(None),) * index.ndim, (None,) * len(self.refinement_shape)
+        c = np.mgrid[tuple(slice(0, sz) for sz in self.refinement_shape)].astype(dtp)
+        c_bc = (
+            (slice(None),)
+            + (None,) * (index.ndim - 1)
+            + (slice(None),) * len(self.refinement_shape)
         )
+        return rs[rs_bc] + c[c_bc]
 
-    def neighborhood(self, index, window_size: int):
+    def neighborhood(self, index, window_size: int, ensemble_axis=None):
         index = self._parse_index(index)
         dtp = np.result_type(index)
+        # TODO
         return (
             np.add.outer(index, np.arange(window_size, dtype=dtp) - window_size // 2)
             % self.size
         )
 
     def parent(self, index):
-        if self.parent_refinement_size is None:
+        if self.parent_refinement_shape is None:
             raise IndexError("you are alone in this world")
-        return index // self.parent_refinement_size
+        index = self._parse_index(index)
+        bc = (slice(None),) + (np.newaxis,) * (index.ndim - 1)
+        return index // self.parent_refinement_shape[bc]
+
+    def index2coord(self, index, **kwargs):
+        return NotImplementedError()
+
+    def coord2index(self, coord, **kwargs):
+        return NotImplementedError()
+
+    def index2volume(self, index, **kwargs):
+        return NotImplementedError()
 
 
-class GridAxis:
+class Grid:
     """Dense (single) axis with periodic boundary conditions.
 
[…]
