### 2024-03-13 e6c01ef
multi_grid/indexing.py: Adapt grids to new structure
