+++ src/re/multi_grid/indexing.py
@@ -3,23 +3,22 @@
 import operator
 from dataclasses import dataclass
 from functools import reduce
-from typing import Any, Callable, Iterable, Optional, Union
+from typing import Callable, Optional, Iterable
 
 import numpy as np
+import numpy.typing as npt
 
-_kind_of_int = Union[tuple[int], list[int], int]
 
-
-@dataclass(kw_only=True)
+@dataclass()
 class GridAtLevel:
-    shape: tuple[int]
-    refinement_shape: tuple[int]
-    parent_refinement_shape: Optional[tuple[int]] = None
+    shape: npt.NDArray[np.int_]
+    splits: npt.NDArray[np.int_]
+    parent_splits: Optional[npt.NDArray[np.int_]] = None
 
-    def __init__(self, shape, refinement_shape, parent_refinement_shape):
-        self.shape = np.asarray(shape)
-        self.refinement_shape = np.asarray(refinement_shape)
-        self.parent_refinement_shape = np.asarray(parent_refinement_shape)
+    def __init__(self, shape, splits, parent_splits):
+        self.shape = np.atleast_1d(shape)
+        self.splits = np.atleast_1d(splits)
+        self.parent_splits = np.atleast_1d(parent_splits)
 
     def _parse_index(self, index):
         index = np.asarray(index)
@@ -31,36 +30,46 @@ class GridAtLevel:
 
     @property
     def size(self):
+        return reduce(operator.mul, self.shape, 1)
+
+    @property
+    def ndim(self):
         return len(self.shape)
 
     def children(self, index) -> np.ndarray:
         index = self._parse_index(index)
         dtp = np.result_type(index)
-        rs = self.refinement_shape[(slice(None),) + (None,) * (index.ndim - 1)]
-        rs_bc = (slice(None),) * index.ndim, (None,) * len(self.refinement_shape)
-        c = np.mgrid[tuple(slice(0, sz) for sz in self.refinement_shape)].astype(dtp)
+        f = self.splits[(slice(None),) + (np.newaxis,) * (index.ndim - 1)]
+        c = np.mgrid[tuple(slice(sz) for sz in self.splits)].astype(dtp)
         c_bc = (
             (slice(None),)
-            + (None,) * (index.ndim - 1)
-            + (slice(None),) * len(self.refinement_shape)
+            + (np.newaxis,) * (index.ndim - 1)
+            + (slice(None),) * self.ndim
         )
-        return rs[rs_bc] + c[c_bc]
+        return (index * f)[..., (np.newaxis,) * self.ndim] + c[c_bc]
 
-    def neighborhood(self, index, window_size: int, ensemble_axis=None):
+    def neighborhood(self, index, window_size: Iterable[int], ensemble_axis=None):
         index = self._parse_index(index)
         dtp = np.result_type(index)
-        # TODO
-        return (
-            np.add.outer(index, np.arange(window_size, dtype=dtp) - window_size // 2)
-            % self.size
+        if ensemble_axis is not None:
+            raise NotImplementedError()
+        window_size = np.asarray(window_size)
+        c = np.mgrid[tuple(slice(sz) for sz in window_size)].astype(dtp)
+        c -= (window_size // 2)[:, (np.newaxis,) * self.ndim]
+        c_bc = (
+            (slice(None),)
+            + (np.newaxis,) * (index.ndim - 1)
+            + (slice(None),) * self.ndim
         )
+        m_bc = (slice(None),) + (np.newaxis,) * (index.ndim - 1 + self.ndim)
+        return (index[..., (np.newaxis,) * self.ndim] + c[c_bc]) % self.shape[m_bc]
 
     def parent(self, index):
-        if self.parent_refinement_shape is None:
[…]
