### 2024-03-13 25a9251
multi_grid/indexing.py: Implement index2coord for HP and RG
