### 2024-03-13 b48bada
Grid.__doc__: Tweak
