### 2024-03-14 1366733
CI: Install dependencies via docker to save time
