+++ src/plot.py
@@ -527,8 +527,6 @@ def _plotHist(f, ax, **kwargs):
 
 def _plot(f, ax, **kwargs):
     _register_cmaps()
-    if isinstance(f, (Field, EnergyHistory)):
-        f = [f]
     if isinstance(f[0], EnergyHistory):
         _plot_history(f, ax, **kwargs)
         return
