### 2024-03-14 ca18e8a
re.multi_grid.indexing: Add FlatGridAtLevel
