### 2024-03-15 93bd19a
re.multi_grid.indexing: flesh out SparseGrid
