+++ src/re/multi_grid/indexing.py
@@ -528,49 +528,128 @@ class FlatGridAtLevel(GridAtLevel):
         return self.index_to_flatindex(window, -1)
 
     def index2coord(self, index, **kwargs):
+        index = self._parse_index(index)
         index = self.flatindex_to_index(index)
         return self.gridAtLevel.index2coord(index, **kwargs)
 
     def coord2index(self, coord, **kwargs):
+        index = self._parse_index(index)
         index = self.gridAtLevel.coord2index(coord, **kwargs)
         return self.index_to_flatindex(index)
 
     def index2volume(self, index, **kwargs):
+        index = self._parse_index(index)
         index = self.flatindex_to_index(index)
         return self.gridAtLevel.index2volume(index, **kwargs)
 
 
-class FlatGrid(Grid):
-    """Same as :class:`Grid` but with a single global integer index for each voxel."""
-    grid: Grid
-    nest: bool
-
-    def __init__(self, grid: Grid, nest=True):
-        self.grid = grid
-        self.nest = nest
-        super().__init__(
-            shape0=grid.shape0, splits=grid.splits, atLevel=FlatGridAtLevel
-        )
-
-    def amend(self, splits, **kwargs):
-        grid = self.grid.amend(splits, **kwargs)
-        return self.__class__(grid, nest=self.nest)
-
-    def at(self, level: int):
-        level = self._parse_level(level)
-
-
-
-
 class SparseGridAtLevel(FlatGridAtLevel):
-    real2flat: np.ndarray
+    real2flat: npt.NDArray[np.int_]
     flat2real: dict
 
+    def __init__(self,
+                 gridAtLevel,
+                 shape0,
+                 all_parent_splits,
+                 real2flat,
+                 flat2real,
+                 ordering='nest'):
+        if not isinstance(gridAtLevel, GridAtLevel):
+            raise TypeError(f"Grid {gridAtLevel.__name__} of invalid type")
+        real2flat = np.asarray(real2flat)
+        if not isinstance(flat2real, dict):
+            raise ValueError("flat2real must be of type dict")
+        if len(flat2real) != real2flat.size:
+            raise ValueError("Index mappings of incompatible size")
+        self.real2flat = real2flat
+        self.flat2real = flat2real
+        super().__init__(
+            gridAtLevel=gridAtLevel,
+            shape0=shape0,
+            all_parent_splits=all_parent_splits,
+            ordering=ordering
+        )
 
-class SparseGrid:
+    def get_flat_index(self, index):
+        return self.real2flat[index]
+
+    def get_real_index(self, index):
+        shp = index.shape
+        return np.array(self.flat2real[ii] for ii in index.ravel()).reshape(shp)
+
+    def _parse_index(self, index):
+        index = self.get_flat_index(index)
+        return super()._parse_index(index)
+
+class SparseGrid(Grid):
     """Realized :class:`FlatGridAtLevel` keeping track of the indices that are actually
     being modeled at the end of the day. This class is especially convenient for
     open boundary conditions but works for arbitrarily sparsely resolved grids."""
 
-    real2flat: tuple[np.ndarray]
-    flat2real: dict
+    real2flat: tuple[npt.NDArray[np.int_]]
+    flat2real: tuple[dict]
+
+    def __init__(self, grid, real2flat, flat2real = None, ordering='nest'):
+        if not isinstance(grid, Grid):
+            raise TypeError(f"Grid {grid.__name__} of invalid type")
+        self.grid = grid
+        self.ordering = ordering
+
+        real2flat = (real2flat, ) if not isinstance(real2flat, tuple) else real2flat
+        real2flat = tuple(np.atleast_1d(r) for r in real2flat)
+        if flat2real is None:
+            flat2real = flat2real = tuple(
[…]
