### 2024-03-15 473febc
re.multi_grid.indexing: flesh out FlatGrid
