+++ src/re/multi_grid/indexing.py
@@ -411,11 +411,11 @@ class FlatGridAtLevel(GridAtLevel):
     """Same as :class:`GridAtLevel` but with a single global integer index for each voxel."""
 
     gridAtLevel: GridAtLevel
-    shape0: npt.NDArray[np.int_]
-    all_parent_splits: tuple[npt.NDArray[np.int_]]
+    gridshape0: npt.NDArray[np.int_]
+    grid_all_splits: tuple[npt.NDArray[np.int_]]
     ordering: str
 
-    def __init__(self, gridAtLevel, shape0, all_parent_splits, ordering='nest'):
+    def __init__(self, gridAtLevel, gridshape0, grid_all_splits, ordering='nest'):
         if not isinstance(gridAtLevel, GridAtLevel):
             raise TypeError(f"Grid {gridAtLevel.__name__} of invalid type")
         self.gridAtLevel = gridAtLevel
@@ -423,17 +423,18 @@ class FlatGridAtLevel(GridAtLevel):
         if ordering not in ['serial', 'nest']:
             raise ValueError(f"Unknown flat index ordering scheme {ordering}")
 
-        self.shape0 = np.asarray(shape0)
-        self.all_parent_splits = tuple(
-            np.atleast_1d(s) for s in all_parent_splits
+        self.gridshape0 = np.asarray(gridshape0)
+        self.grid_all_splits = tuple(
+            np.atleast_1d(s) for s in grid_all_splits
         )
         super().__init__(
-            self.gridAtLevel.shape,
-            self.gridAtLevel.splits,
-            self.gridAtLevel.parent_splits
+            shape=(reduce(operator.mul, gridAtLevel.shape, 1), ),
+            splits=None,
+            parent_splits=None
         )
 
     def _parse_index(self, index):
+        #TODO think about adding dummy axis to flat indices to match Grids
         index = np.asarray(index)
         if np.any(np.abs(index) >= self.size):
             nm = self.__class__.__name__
@@ -442,24 +443,27 @@ class FlatGridAtLevel(GridAtLevel):
         return index % self.size
 
     def _weights_serial(self, levelshift):
-        shape = self.shape
-        if levelshift == 1:
-            shape *= self.splits
+        shape = (self.gridshape0, ) + self.grid_all_splits
+        if levelshift == 0:
+            shape = shape[:-1]
         elif levelshift == -1:
-            shape //= self.parent_splits
+            shape = shape[:-2]
         else:
-            raise ValueError(f"Inconsistent shift in level: {levelshift}")
+            if levelshift != 1:
+                raise ValueError(f"Inconsistent shift in level: {levelshift}")
+        shape = reduce(operator.mul, shape)
         wgt = np.append(shape[1:], 1)
         return np.cumprod(wgt[::-1])[::-1]
 
     def _weights_nest(self, levelshift):
-        wgts = (self.shape0, ) + self.all_parent_splits
-        if levelshift == 1:
-            wgts += (self.splits, )
-        elif levelshift == -1:
+        wgts = (self.gridshape0, ) + self.grid_all_splits
+        if levelshift == 0:
             wgts = wgts[:-1]
+        elif levelshift == -1:
+            wgts = wgts[:-2]
         else:
-            raise ValueError(f"Inconsistent shift in level: {levelshift}")
+            if levelshift != 1:
+                raise ValueError(f"Inconsistent shift in level: {levelshift}")
         return np.stack(wgts, axis=0)
 
     def index_to_flatindex(self, index, levelshift = 0):
@@ -575,45 +579,64 @@ class FlatGrid(Grid):
                             ordering=self.ordering
         )
 
-# TODO still work in progress...
[…]
