### 2024-03-18 6620b59
multi_grid/indexing.py: Format
