+++ src/re/multi_grid/indexing.py
@@ -415,26 +415,24 @@ class FlatGridAtLevel(GridAtLevel):
     grid_all_splits: tuple[npt.NDArray[np.int_]]
     ordering: str
 
-    def __init__(self, gridAtLevel, gridshape0, grid_all_splits, ordering='nest'):
+    def __init__(self, gridAtLevel, gridshape0, grid_all_splits, ordering="nest"):
         if not isinstance(gridAtLevel, GridAtLevel):
             raise TypeError(f"Grid {gridAtLevel.__name__} of invalid type")
         self.gridAtLevel = gridAtLevel
         self.ordering = ordering
-        if ordering not in ['serial', 'nest']:
+        if ordering not in ["serial", "nest"]:
             raise ValueError(f"Unknown flat index ordering scheme {ordering}")
 
         self.gridshape0 = np.asarray(gridshape0)
-        self.grid_all_splits = tuple(
-            np.atleast_1d(s) for s in grid_all_splits
-        )
+        self.grid_all_splits = tuple(np.atleast_1d(s) for s in grid_all_splits)
         super().__init__(
-            shape=(reduce(operator.mul, gridAtLevel.shape, 1), ),
+            shape=(reduce(operator.mul, gridAtLevel.shape, 1),),
             splits=None,
-            parent_splits=None
+            parent_splits=None,
         )
 
     def _parse_index(self, index):
-        #TODO think about adding dummy axis to flat indices to match Grids
+        # TODO think about adding dummy axis to flat indices to match Grids
         index = np.asarray(index)
         if np.any(np.abs(index) >= self.size):
             nm = self.__class__.__name__
@@ -443,7 +441,7 @@ class FlatGridAtLevel(GridAtLevel):
         return index % self.size
 
     def _weights_serial(self, levelshift):
-        shape = (self.gridshape0, ) + self.grid_all_splits
+        shape = (self.gridshape0,) + self.grid_all_splits
         if levelshift == 0:
             shape = shape[:-1]
         elif levelshift == -1:
@@ -456,7 +454,7 @@ class FlatGridAtLevel(GridAtLevel):
         return np.cumprod(wgt[::-1])[::-1]
 
     def _weights_nest(self, levelshift):
-        wgts = (self.gridshape0, ) + self.grid_all_splits
+        wgts = (self.gridshape0,) + self.grid_all_splits
         if levelshift == 0:
             wgts = wgts[:-1]
         elif levelshift == -1:
@@ -466,28 +464,28 @@ class FlatGridAtLevel(GridAtLevel):
                 raise ValueError(f"Inconsistent shift in level: {levelshift}")
         return np.stack(wgts, axis=0)
 
-    def index_to_flatindex(self, index, levelshift = 0):
-        if self.ordering == 'serial':
+    def index_to_flatindex(self, index, levelshift=0):
+        if self.ordering == "serial":
             wgt = self._weights_serial(levelshift)
-            wgt = wgt[(slice(None), ) + (np.newaxis,) * (index.ndim - 1)]
-            return (wgt * index).sum(axis = 0).astype(index)
-        if self.ordering == 'nest':
-            #TODO vectorize better
+            wgt = wgt[(slice(None),) + (np.newaxis,) * (index.ndim - 1)]
+            return (wgt * index).sum(axis=0).astype(index)
+        if self.ordering == "nest":
+            # TODO vectorize better
             fid = np.zeros(index.shape[1:], dtype=index.dtype)
             wgts = self._weights_nest(levelshift)
             for n, ww in enumerate(wgts):
                 j = 0
                 for ax in range(ww.size):
                     j *= ww[ax]
-                    j += (index[ax] // wgts[(n+1):, ax].prod()) % ww[ax]
+                    j += (index[ax] // wgts[(n + 1) :, ax].prod()) % ww[ax]
[…]
