### 2024-03-18 0a67d9c
multi_grid/indexing.py: Implement OGridAtLevel
