### 2024-03-18 7c70e0a
370-implement-evidence-calculation: address review.
