+++ src/re/evidence_lower_bound.py
@@ -4,9 +4,10 @@ import scipy.linalg as slg
 import scipy.sparse.linalg as ssl
 
 from .evi import Samples
-from .likelihood import StandardHamiltonian
+from .likelihood import Likelihood
+from .optimize_kl import _StandardHamiltonian as StandardHamiltonian
 from .logger import logger
-from .tree_math import tree_shape
+from .tree_math.vector_math import size
 
 
 class _Projector(ssl.LinearOperator):
@@ -24,42 +25,40 @@ class _Projector(ssl.LinearOperator):
         Operator representing the projection.
     """
     def __init__(self, eigenvectors):
-        super().__init__(np.dtype('f8'), 2 * (eigenvectors.shape[0],))
+        super().__init__(eigenvectors.dtype, 2 * (eigenvectors.shape[0],))
         self.eigenvectors = eigenvectors
 
     def _matvec(self, x):
         res = x.copy()
         for eigenvector in self.eigenvectors.T:
-            res -= eigenvector * eigenvector.dot(x)
+            res -= eigenvector * np.vdot(eigenvector, x)
         return res
 
     def _rmatvec(self, x):
         return self._matvec(x)
 
-
 def _explicify(M):
-    identity = np.identity(M.shape[0], dtype=np.float64)
+    n = M.shape[0]
     m = []
-    for v in identity:
-        m.append(M.matvec(v))
-    return np.vstack(m).T
+    for i in range(n):
+        basis_vector = np.zeros(n)
+        basis_vector[i] = 1
+        m.append(M @ basis_vector)
+    return np.stack(m, axis=1)
 
 
-def _ravel_metric(metric, position):
-    dim = 0
-    shapes = tree_shape(metric(position, position).tree)
-    for s in shapes.values():
-        dim += np.prod(s)
-    shape = 2*(int(dim),)
+def _ravel_metric(metric, position, dtype):
+    shape = 2*(size(metric(position, position)),)
 
     ravel = lambda x: jax.flatten_util.ravel_pytree(x)[0]
     unravel = lambda x: jax.linear_transpose(ravel, position)(x)[0]
     met = lambda x: ravel(metric(position, unravel(x)))
 
-    return ssl.LinearOperator(shape=shape, matvec=met)
+    return ssl.LinearOperator(shape=shape, dtype=dtype, matvec=met)
 
 
-def _eigsh(metric, n_eigenvalues, tot_dofs, min_lh_eval=1e-4, batch_number=10, tol=0., verbose=True):
+def _eigsh(metric, n_eigenvalues, tot_dofs, min_lh_eval=1e-4, batch_size=10, tol=0.,
+           verbose=True):
     metric_size = metric.shape[0]
     eigenvectors = None
     if n_eigenvalues > tot_dofs:
@@ -76,9 +75,9 @@ def _eigsh(metric, n_eigenvalues, tot_dofs, min_lh_eval=1e-4, batch_number=10, t
         eigenvalues = np.flip(eigenvalues)
     else:
         # Set up batches
-        batch_size = n_eigenvalues // batch_number
-        batches = [batch_size, ] * (batch_number - 1)
-        batches += [n_eigenvalues - batch_size * (batch_number - 1), ]
+        batch_size = n_eigenvalues // batch_size
+        batches = [batch_size, ] * (batch_size - 1)
+        batches += [n_eigenvalues - batch_size * (batch_size - 1), ]
         eigenvalues, projected_metric = None, metric
         for batch in batches:
             if verbose:
@@ -86,8 +85,8 @@ def _eigsh(metric, n_eigenvalues, tot_dofs, min_lh_eval=1e-4, batch_number=10, t
             # Get eigensystem for current batch
[…]
