### 2024-03-21 103fc73
HEALPixGrid*: Make fill_strategy class property
