### 2024-03-21 dc2ebbe
re.multi_grid: update SparseGrid
