### 2024-03-22 3a20021
re.multigrid.indexing: add len 1 axis to FlatGridAtLevel and fixup mapping
