### 2024-03-22 45e7c95
re.multi_grid.indexing: update parse_index
