### 2024-03-24 70c03d3
re.multi_grid.indexing: SparseGrid indexing and valid neighbors
