### 2024-03-24 c9fe2f7
370-implement-evidence-calculation: better docstrings, improved readability.
