### 2024-03-24 82ac740
SparseGrid: Check unique sorted indexing
