+++ src/re/evidence_lower_bound.py
@@ -5,8 +5,8 @@ import scipy.sparse.linalg as ssl
 
 from .evi import Samples
 from .likelihood import Likelihood
-from .optimize_kl import _StandardHamiltonian as StandardHamiltonian
 from .logger import logger
+from .optimize_kl import _StandardHamiltonian as StandardHamiltonian
 from .tree_math.vector_math import size
 
 
@@ -24,6 +24,7 @@ class _Projector(ssl.LinearOperator):
     Projector : LinearOperator
         Operator representing the projection.
     """
+
     def __init__(self, eigenvectors):
         super().__init__(eigenvectors.dtype, 2 * (eigenvectors.shape[0],))
         self.eigenvectors = eigenvectors
@@ -49,7 +50,7 @@ def _explicify(M):
 
 
 def _ravel_metric(metric, position, dtype):
-    shape = 2*(size(metric(position, position)),)
+    shape = 2 * (size(metric(position, position)),)
 
     ravel = lambda x: jax.flatten_util.ravel_pytree(x)[0]
     unravel = lambda x: jax.linear_transpose(ravel, position)(x)[0]
@@ -58,38 +59,60 @@ def _ravel_metric(metric, position, dtype):
     return ssl.LinearOperator(shape=shape, dtype=dtype, matvec=met)
 
 
-def _eigsh(metric, n_eigenvalues, tot_dofs, min_lh_eval=1e-4, batch_size=10, tol=0.,
-           verbose=True):
+def _eigsh(
+    metric,
+    n_eigenvalues,
+    tot_dofs,
+    min_lh_eval=1e-4,
+    batch_size=10,
+    tol=0.0,
+    verbose=True,
+):
     metric_size = metric.shape[0]
     eigenvectors = None
     if n_eigenvalues > tot_dofs:
-        raise ValueError("Number of requested eigenvalues "
-                         "exceeds the number of relevant degrees of freedom!")
+        raise ValueError(
+            "Number of requested eigenvalues "
+            "exceeds the number of relevant degrees of freedom!"
+        )
 
     if tot_dofs == n_eigenvalues:
         # Compute exact eigensystem
         if verbose:
             logger.info(f"Computing all {tot_dofs} relevant metric eigenvalues.")
-        eigenvalues = slg.eigh(_explicify(metric),
-                               eigvals_only=True,
-                               subset_by_index=[metric_size - tot_dofs, metric_size - 1])
+        eigenvalues = slg.eigh(
+            _explicify(metric),
+            eigvals_only=True,
+            subset_by_index=[metric_size - tot_dofs, metric_size - 1],
+        )
         eigenvalues = np.flip(eigenvalues)
     else:
         # Set up batches
         batch_size = n_eigenvalues // batch_size
-        batches = [batch_size, ] * (batch_size - 1)
-        batches += [n_eigenvalues - batch_size * (batch_size - 1), ]
+        batches = [
+            batch_size,
+        ] * (batch_size - 1)
+        batches += [
+            n_eigenvalues - batch_size * (batch_size - 1),
+        ]
         eigenvalues, projected_metric = None, metric
         for batch in batches:
             if verbose:
                 logger.info(f"\nNumber of eigenvalues being computed: {batch}")
             # Get eigensystem for current batch
-            eigvals, eigvecs = ssl.eigsh(projected_metric, k=batch, tol=tol,
-                                         return_eigenvectors=True, which='LM')
+            eigvals, eigvecs = ssl.eigsh(
[…]
