+++ test/test_re/test_estimate_evidence_lower_bound.py
@@ -14,14 +14,13 @@
 # Copyright(C) 2024 Max-Planck-Society
 # Author: Matteo Guardiani
 
+import jax
+import jax.random as random
+import numpy as np
+import pytest
+
 import nifty8 as ift
 import nifty8.re as jft
-import numpy as np
-import jax
-import pytest
-import jax.random as random
-
-from nifty8.re import ShapeWithDtype
 
 pmp = pytest.mark.parametrize
 
@@ -45,7 +44,7 @@ def get_linear_response(slope_op, intercept_op, sampling_points):
     return jft.Model(response, domain=slope_op.domain | intercept_op.domain)
 
 
-@pmp('seed', [0, 42])
+@pmp("seed", [0, 42])
 def test_estimate_evidence_lower_bound(seed):
     # Set up signal
     n_datapoints = 8
@@ -53,16 +52,16 @@ def test_estimate_evidence_lower_bound(seed):
     key = random.PRNGKey(seed)
     key, subkey = random.split(key)
 
-    q = 0.
+    q = 0.0
     m_slope = 1.5
 
     sigma_q = 1.5
     sigma_m = 1.8
 
-    intercept = jft.NormalPrior(0., sigma_q, name="intercept")
-    slope = jft.NormalPrior(0., sigma_m, name="slope")
+    intercept = jft.NormalPrior(0.0, sigma_q, name="intercept")
+    slope = jft.NormalPrior(0.0, sigma_m, name="slope")
 
-    x = random.uniform(subkey, shape=(n_datapoints,), minval=0., maxval=10.)
+    x = random.uniform(subkey, shape=(n_datapoints,), minval=0.0, maxval=10.0)
     y = q + m_slope * x
 
     linear_response = get_linear_response(slope, intercept, x)
@@ -75,7 +74,7 @@ def test_estimate_evidence_lower_bound(seed):
     noise = random.normal(subkey, shape=(n_datapoints,)) * noise_level
     data = y + noise
 
-    N_inv = np.diag(np.ones_like(data) * (1. / noise_level ** 2))
+    N_inv = np.diag(np.ones_like(data) * (1.0 / noise_level**2))
 
     S = np.identity(2)
     S_inv = np.identity(2)
@@ -90,14 +89,20 @@ def test_estimate_evidence_lower_bound(seed):
     det_2pi_D = np.linalg.det(2 * np.pi * D)
     det_2pi_S = np.linalg.det(2 * np.pi * S)
 
-    H_0 = 0.5 * (np.vdot(data, (N_inv @ data)) + n_datapoints * np.log(2 * np.pi * noise_level ** 2)
-                 + np.log(det_2pi_S) - m_dag_j)
+    H_0 = 0.5 * (
+        np.vdot(data, (N_inv @ data))
+        + n_datapoints * np.log(2 * np.pi * noise_level**2)
+        + np.log(det_2pi_S)
+        - m_dag_j
+    )
 
     evidence = -H_0 + 0.5 * np.log(det_2pi_D)
-    nifty_adjusted_evidence = evidence + 0.5 * n_datapoints * np.log(2 * np.pi * noise_level ** 2)
-    likelihood_energy = jft.Gaussian(data=data,
-                                     noise_cov_inv=lambda x: (1. / noise_level ** 2) * x).amend(
-        linear_response)
+    nifty_adjusted_evidence = evidence + 0.5 * n_datapoints * np.log(
+        2 * np.pi * noise_level**2
+    )
+    likelihood_energy = jft.Gaussian(
+        data=data, noise_cov_inv=lambda x: (1.0 / noise_level**2) * x
+    ).amend(linear_response)
 
     # Minimization parameters
     n_iterations = 4
[…]
