+++ src/re/evidence_lower_bound.py
@@ -1,3 +1,7 @@
+#!/usr/bin/env python3
+
+# SPDX-License-Identifier: GPL-2.0+ OR BSD-2-Clause
+
 import jax.flatten_util
 import numpy as np
 import scipy.linalg as slg
