### 2024-04-03 9e0f278
OptimizeVI: Tweak `sample_mode` description
