### 2024-04-04 11f3a2a
OGridAtLevel.neighborhood: Fix output assignment
