### 2024-04-08 #6382 @dfm
:wave: @williamjameshandley, @thomaspinder, @apizzuto — Would any of you be available and willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
