### 2024-04-09 164aeda
Add a switch for the proper Hartley convention
