### 2024-04-09 d8a513f
config.update: Disallow any other keys except for hartley
