### 2024-04-09 bdf6255
config.update: Add doc-string
