### 2024-04-09 325d311
Undeprecate remaining tree_map calls
