### 2024-04-10 aca83af
config.update: Fix spelling mistake
