### 2024-04-28 NIFTy-PPL/NIFTy#12 @apizzuto
Remind user to install extra `nifty[re]` dependencies when installing editable version
Part of https://github.com/openjournals/joss-reviews/issues/6593

I decided to install the editable version in a fresh virtual environment and if following the directions in the README, this will leave the user without the extra dependencies that are installed when installing from pypi with the `[re]` specification. If possible, please just list the extra dependencies the user should install in the README when describing how to install the editable version (eg. `jax` and `jaxlib` based off of the `setup.py` file)
