### 2024-04-28 NIFTy-PPL/NIFTy#13 @apizzuto
Some modules not installed when running tests
ModuleImport errors when running tests as suggested in the README.

```
========================================== short test summary info ===========================================
ERROR test/test_mpi/test_kl.py
ERROR test/test_mpi/test_sample_list.py
ERROR test/test_mpi/test_sync.py
ERROR test/test_numpy_accuracy.py
ERROR test/test_optimize_kl_out.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 5 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
================================== 2 skipped, 4 warnings, 5 errors in 7.93s ==================================
```

Two tests fail from `h5py` or `ducc0`, which are easily installed via `pip`. The others fail because of `mpi4py` which requires an MPI installation.

Part of https://github.com/openjournals/joss-reviews/issues/6593#issuecomment-2045195836
