### 2024-04-28 NIFTy-PPL/NIFTy#14 @apizzuto
Broken link in README
There is a broken link that 404s for me in the README where it reads `For a quick start, you can browse through the [informal introduction](https://ift.pages.mpcdf.de/nifty/user/code.html) or dive into NIFTy by running one of the demonstrations, e.g.:`

The link: https://ift.pages.mpcdf.de/nifty/user/code.html

Part of https://github.com/openjournals/joss-reviews/issues/6593
