### 2024-04-28 NIFTy-PPL/NIFTy#15 @apizzuto
Please add contribution instructions
As part of https://github.com/openjournals/joss-reviews/issues/6593:

The review asks: `Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support`. You all list (2) and (3) but it was not clear to me if there were contribution instructions
