### 2024-04-28 NIFTy-PPL/NIFTy#16 @apizzuto
Plot requires labels in `demos/0_intro.py`
Part of https://github.com/openjournals/joss-reviews/issues/6593

Just a tiny nit, but can you please label the `Ax1` subplot in the `0_intro.py` demo? Looks like the values correspond to `(cfm.amplitude(pos_truth)[1:], post_a_mean)` but this isn't obvious when looking at the figure.
