### 2024-04-28 #6593 @apizzuto
Hello authors — thank you for a lovely piece of software! I enjoyed installing and playing around with this tool. Overall, I found the paper to be extremely clear, the API docs and installation directions were well done, and the tutorials were a really nice introduction to solving real problems with the software.

I left a couple of tiny nits as issues in the repository, they are pretty much all documentation based, so hopefully they won't be hard to address. The only remaining technical parts of my review involve reproducing plots from the paper, which I will try to do in the next few days. I will only benchmark a few performance metrics just to make sure that things seem to line up with your performance plot. I will circle back once I've completed those parts (and I'll check off the remaining boxes in my review once the issues I opened up are addressed). Cheers!
