### 2024-05-10 a210041
README.md: Install re dependency also during editable install
Resolves #12 .
