### 2024-05-10 66fd9b5
README.md: Install re dependency also during editable install
Resolves #12 .
