### 2024-05-10 e1ed617
pyproject.toml: Remove redundant scipy dependency
