### 2024-05-10 7f551c4
README.md: Fix test instructions
Resolves #13 .
