### 2024-05-10 1786ce7
README.md: Fix test instructions
Resolves #13 .
