+++ README.md
@@ -171,11 +171,13 @@ Note: Make sure that you reinstall nifty after each change since sphinx imports 
 
 ### Run the tests
 
-To run the tests, install the `test` extra requirements and afterwards run pytest (and create a coverage report) via
+To run the tests, install all optional requirements `'nifty8[all]'` and afterwards run pytest (and create a coverage report) via
 
 ```
 pytest --cov=nifty8 test
 ```
+
+If you are writing your own tests, it is often sufficient to just install the optional test dependencies `'nifty8[test]'`. However, to run the full test suit including tests of optional functionality, it is assumed that all optional dependencies are installed.
 
 ## First Steps
 
