### 2024-05-10 NIFTy-PPL/NIFTy#14 @Edenhofer
That's bad :|  #19 should fix this. Thanks for the report!
