### 2024-05-10 NIFTy-PPL/NIFTy#15 @Edenhofer
Thanks for filing this issue. All kinds of contributions are welcome and we now provide a brief description on how to get started.
