+++ src/library/correlated_fields_simple.py
@@ -168,3 +168,142 @@ def SimpleCorrelatedField(
         #     warn(f"no JAX operator for this configuration;\n{e}")
 
     return op
+
+
+def ModifiedSimpleCorrelatedField(
+    target,
+    offset_mean,
+    offset_std,
+    fluctuations,
+    flexibility,
+    asperity,
+    loglogavgslope,
+    prefix="",
+    harmonic_partner=None,
+    power_space_model_builder=None,
+):
+    """Simplified version of :class:`~nifty8.library.correlated_fields.CorrelatedFieldMaker`.
+
+    Assumes `total_N = 0`, `dofdex = None` and the presence of only one power
+    spectrum, i.e. only one call of
+    :func:`~nifty8.library.correlated_fields.CorrelatedFieldMaker.add_fluctuations`.
+
+    See also
+    --------
+    * The simple correlated field model has first been described in "Comparison
+      of classical and Bayesian imaging in radio interferometry", A&A 646, A84
+      (2021) by P. Arras et al.
+      `<https://doi.org/10.1051/0004-6361/202039258>`_
+
+    Consider citing this paper, if you use the simple correlated field model.
+    """
+    target = DomainTuple.make(target)
+    if len(target) != 1:
+        raise ValueError
+    target = target[0]
+    if harmonic_partner is None:
+        harmonic_partner = target.get_default_codomain()
+    else:
+        target.check_codomain(harmonic_partner)
+        harmonic_partner.check_codomain(target)
+    for kk in (fluctuations, loglogavgslope):
+        if len(kk) != 2:
+            raise TypeError
+    for kk in (offset_std, flexibility, asperity):
+        if not (kk is None or len(kk) == 2):
+            raise TypeError
+    if flexibility is None and asperity is not None:
+        raise ValueError
+    fluct = LognormalTransform(*fluctuations, prefix + 'fluctuations', 0)
+    avgsl = NormalTransform(*loglogavgslope, prefix + 'loglogavgslope', 0)
+
+    pspace = PowerSpace(harmonic_partner)
+    twolog = _TwoLogIntegrations(pspace)
+    expander = ContractionOperator(twolog.domain, 0).adjoint
+    ps_expander = ContractionOperator(pspace, 0).adjoint
+    vslope = makeOp(makeField(pspace, _relative_log_k_lengths(pspace)))
+    slope = vslope @ ps_expander @ avgsl
+    a = slope
+
+    if flexibility is not None:
+        flex = LognormalTransform(*flexibility, prefix + 'flexibility', 0)
+        dom = twolog.domain[0]
+        vflex = np.empty(dom.shape)
+        vflex[0] = vflex[1] = np.sqrt(_log_vol(pspace))
+        vflex = makeOp(makeField(dom, vflex))
+        sig_flex = vflex @ expander @ flex
+        xi = ducktape(dom, None, prefix + 'spectrum')
+
+        shift = np.empty(dom.shape)
+        shift[0] = _log_vol(pspace)**2 / 12.
+        shift[1] = 1
+        shift = makeField(dom, shift)
+        if asperity is None:
+            asp = makeOp(shift.ptw("sqrt")) @ (xi*sig_flex)
+        else:
+            asp = LognormalTransform(*asperity, prefix + 'asperity', 0)
+            vasp = np.empty(dom.shape)
+            vasp[0] = 1
+            vasp[1] = 0
+            vasp = makeOp(makeField(dom, vasp))
+            sig_asp = vasp @ expander @ asp
+            asp = xi*sig_flex*(Adder(shift) @ sig_asp).ptw("sqrt")
+        a = a + _SlopeRemover(pspace, 0) @ twolog @ asp
+    a = _Normalization(pspace, 0) @ a
[…]
