### 2024-05-10 6946a87
0_intro.py: Set plot extents and fix spectrum abscissa
Resolves #16 .
