### 2024-05-10 f30eec0
0_intro.py: Set plot extents and fix spectrum abscissa
Resolves #16 .
