### 2024-05-10 NIFTy-PPL/NIFTy#16 @Edenhofer
Thanks! Right, yeah it is definitely not obvious what this plot shows :see_no_evil: #21 adds labels and tweaks the other plots some more too.
