### 2024-05-14 58bfa69
re.multi_grid.indexing: Return validity of neighbors
