+++ src/re/multi_grid/indexing.py
@@ -52,11 +52,9 @@ class GridAtLevel:
         )
         return (index * f)[..., (np.newaxis,) * self.ndim] + c[c_bc]
 
-    def neighborhood(self, index, window_size: Iterable[int], ensemble_axis=None):
+    def neighborhood(self, index, window_size: Iterable[int]):
         index = self._parse_index(index)
         dtp = np.result_type(index)
-        if ensemble_axis is not None:
-            raise NotImplementedError()
         window_size = np.asarray(window_size)
         c = np.mgrid[tuple(slice(sz) for sz in window_size)].astype(dtp)
         c -= (window_size // 2)[:, (np.newaxis,) * self.ndim]
@@ -66,7 +64,8 @@ class GridAtLevel:
             + (slice(None),) * self.ndim
         )
         m_bc = (slice(None),) + (np.newaxis,) * (index.ndim - 1 + self.ndim)
-        return (index[..., (np.newaxis,) * self.ndim] + c[c_bc]) % self.shape[m_bc]
+        res = (index[..., (np.newaxis,) * self.ndim] + c[c_bc]) % self.shape[m_bc]
+        return res, np.ones(res.shape[1:], dtype=bool)
 
     def parent(self, index):
         if self.parent_splits is None:
@@ -221,11 +220,8 @@ class HEALPixGridAtLevel(GridAtLevel):
         self.fill_strategy = fill_strategy.lower()
         super().__init__(shape=size, splits=splits, parent_splits=parent_splits)
 
-    def neighborhood(self, index, window_size: int, ensemble_axis=None):
+    def neighborhood(self, index, window_size: int):
         from healpy.pixelfunc import get_all_neighbours
-
-        if ensemble_axis is not None:
-            raise NotImplementedError()
 
         dtp = np.result_type(index)
         if window_size not in (1, 9, self.size):
@@ -244,6 +240,8 @@ class HEALPixGridAtLevel(GridAtLevel):
         nbr = nbr.reshape(window_size - 1, n_pix).T
         neighbors[:, 1:] = nbr
 
+        valid = neighbors != -1
+
         if self.fill_strategy == "unique":
             neighbors = _fill_bad_healpix_neighbors(neighbors)
         elif self.fill_strategy == "same":
@@ -255,7 +253,7 @@ class HEALPixGridAtLevel(GridAtLevel):
 
         neighbors = np.squeeze(neighbors, axis=0) if index_shape == () else neighbors
         neighbors = neighbors.reshape(index_shape + (window_size,))
-        return neighbors.astype(dtp)[np.newaxis]
+        return neighbors.astype(dtp)[np.newaxis], valid.reshape(neighbors.shape)
 
     def index2coord(self, index, **kwargs):
         from healpy.pixelfunc import pix2vec
@@ -366,17 +364,24 @@ class OGridAtLevel(GridAtLevel):
             else window_size
         )
         assert len(window_size) == self.ngrids
-        neighborhood = tuple(
-            g.neighborhood(index[i], wsz)
-            for i, g, wsz in zip(islice, self.grids, window_size)
-        )
+        neighborhood = []
+        valid = []
+        for i, g, wsz in zip(islice, self.grids, window_size):
+            n, v = g.neighborhood(index[i], wsz)
+            neighborhood.append(n)
+            valid.append(v)
+
         # Make initial entry broadcast-able to the full final shape
         out = neighborhood[0].reshape(
             (slice(None),) * neighborhood[0].ndim
             + (np.newaxis,) * (ndims_sum - self.grids[0].ndim)
         )
+        vout = valid[0].reshape(
+            (slice(None),) * valid[0].ndim
+            + (np.newaxis,) * (ndims_sum - self.grids[0].ndim)
+        )
         # Successively concatenate all broadcasted neighbors
[…]
