### 2024-05-16 d9ec2fd
general model for multifrequency astro sky
