### 2024-05-22 NIFTy-PPL/NIFTy#22 @Abinashbunty
Missing DOIs for references
Hi authors. This issue is part of my ongoing review in openjournals/joss-reviews#6593

When running the checks for the references, the bot throws warnings for the following references. I have gathered the DOIs for most except some software based references. Feel free to have a look into it. 😄

- Physics of the Interstellar and Intergalactic Medium: https://doi.org/10.2307/j.ctvcm4hzr
- The Pan-STARRS1 Surveys: https://doi.org/10.48550/arXiv.1612.05560
- Variational Inference with Normalizing Flows: https://doi.org/10.48550/arXiv.1505.05770
- The No-U-Turn Sampler: Adaptively Setting Path Lengths in Hamiltonian Monte Carlo: https://doi.org/10.48550/arXiv.1111.4246
- Equinox: neural networks in JAX via callable PyTrees and filtered transformations : https://doi.org/10.48550/arXiv.2111.00254
- Efficient and Modular Implicit Differentiation: https://doi.org/10.48550/arXiv.2105.15183
- Scalable Variational Gaussian Process Classification: https://doi.org/10.48550/arXiv.1411.2005
- GPflow: A Gaussian Process Library Using Tensorf: https://doi.org/10.48550/arXiv.1610.08733
- Pyro: Deep Universal Probabilistic Programming: https://doi.org/10.48550/arXiv.1810.09538
- NIFTY: A versatile Python library for signal inference: https://doi.org/10.1051/0004-6361/201321236
- Re-Envisioning Numerical Information Field Theory (NIFTy): A Library for Gaussian Processes and Variational Inference: https://doi.org/10.48550/arXiv.2402.16683
