### 2024-05-23 f8e7670
re.multi_grid: Test Grid and HEALPixGrid and fixups
