+++ src/re/multi_grid/indexing.py
@@ -13,12 +13,16 @@ import numpy.typing as npt
 class GridAtLevel:
     shape: npt.NDArray[np.int_]
     splits: npt.NDArray[np.int_]
-    parent_splits: Optional[npt.NDArray[np.int_]] = None
+    parent_splits: Optional[npt.NDArray[np.int_]]
 
-    def __init__(self, shape, splits, parent_splits):
+    def __init__(self, shape, splits=None, parent_splits=None):
         self.shape = np.atleast_1d(shape)
-        self.splits = np.atleast_1d(splits)
-        self.parent_splits = np.atleast_1d(parent_splits)
+        if splits is not None:
+            splits = np.atleast_1d(splits)
+        self.splits = splits
+        if parent_splits is not None:
+            parent_splits = np.atleast_1d(parent_splits)
+        self.parent_splits = parent_splits
 
     def _parse_index(self, index):
         index = np.asarray(index)
@@ -26,7 +30,7 @@ class GridAtLevel:
             l = index.shape[0]
             ve = f"index {index} is of invalid length {l} for shape {self.shape}"
             raise IndexError(ve)
-        if np.any(np.any(np.abs(idx) >= s) for idx, s in zip(index, self.shape)):
+        if np.any(tuple(np.any(np.abs(idx) >= s) for idx, s in zip(index, self.shape))):
             nm = self.__class__.__name__
             ve = f"index {index} is out of bounds for {nm} with shape {self.shape}"
             raise IndexError(ve)
@@ -41,6 +45,8 @@ class GridAtLevel:
         return len(self.shape)
 
     def children(self, index) -> np.ndarray:
+        if self.splits is None:
+            raise IndexError("This level has no children")
         index = self._parse_index(index)
         dtp = np.result_type(index)
         f = self.splits[(slice(None),) + (np.newaxis,) * (index.ndim - 1)]
@@ -50,21 +56,23 @@ class GridAtLevel:
             + (np.newaxis,) * (index.ndim - 1)
             + (slice(None),) * self.ndim
         )
-        return (index * f)[..., (np.newaxis,) * self.ndim] + c[c_bc]
+        ids = index * f
+        return ids[(slice(None),) * ids.ndim + (np.newaxis,) * self.ndim] + c[c_bc]
 
     def neighborhood(self, index, window_size: Iterable[int]):
         index = self._parse_index(index)
         dtp = np.result_type(index)
         window_size = np.asarray(window_size)
         c = np.mgrid[tuple(slice(sz) for sz in window_size)].astype(dtp)
-        c -= (window_size // 2)[:, (np.newaxis,) * self.ndim]
+        c -= (window_size // 2)[(slice(None),) + (np.newaxis,) * self.ndim]
         c_bc = (
             (slice(None),)
             + (np.newaxis,) * (index.ndim - 1)
             + (slice(None),) * self.ndim
         )
         m_bc = (slice(None),) + (np.newaxis,) * (index.ndim - 1 + self.ndim)
-        res = (index[..., (np.newaxis,) * self.ndim] + c[c_bc]) % self.shape[m_bc]
+        id_bc = (slice(None),) * index.ndim + (np.newaxis,) * self.ndim
+        res = (index[id_bc] + c[c_bc]) % self.shape[m_bc]
         return res, np.ones(res.shape[1:], dtype=bool)
 
     def parent(self, index):
@@ -95,7 +103,7 @@ class Grid:
     atLevel: Callable
 
     def __init__(self, *, shape0, splits, atLevel=GridAtLevel):
-        self.shape0 = np.asarray(shape0)
+        self.shape0 = np.atleast_1d(shape0)
[…]
