### 2024-05-23 b4fd774
re.multi_grid: Test FlatGrid
