### 2024-05-23 24677c7
re.multi_grid: Test FlatGrid
