### 2024-05-23 585721c
re.multi_grid: Test SparseGrid
