### 2024-05-23 3666ae3
re.multi_grid: Test SparseGrid
