### 2024-05-24 NIFTy-PPL/NIFTy#22 @Edenhofer
Many thanks for going through the trouble of finding all the DOIs! I'm currently on vacation without my laptop until the the end of next week but I'll incorporate the DOIs right afterwards.
