### 2024-05-24 56ea65f
qpo: add ModifiedSimpleCorrelatedField to the init
