+++ src/__init__.py
@@ -93,7 +93,7 @@ from .library.adjust_variances import (make_adjust_variances_hamiltonian,
                                        do_adjust_variances)
 from .library.nft import Gridder, FinuFFT, Nufft
 from .library.correlated_fields import CorrelatedFieldMaker
-from .library.correlated_fields_simple import SimpleCorrelatedField
+from .library.correlated_fields_simple import SimpleCorrelatedField, ModifiedSimpleCorrelatedField
 from .library.variational_models import MeanFieldVI, FullCovarianceVI
 
 from . import extra
