### 2024-05-25 92b4965
Tiny fixes for JOSS review
