+++ paper/paper.bib
@@ -21,7 +21,7 @@
 @article{Popescu2002,
   author        = {{Popescu}, Cristina C. and {Tuffs}, Richard J.},
   title         = {{The percentage of stellar light re-radiated by dust in late-type Virgo Cluster galaxies}},
-  journal       = {\mnras},
+  journal       = {Monthly Notices of the RAS},
   keywords      = {galaxies: clusters: individual: Virgo Cluster, galaxies: fundamental parameters, galaxies: photometry, galaxies: spiral, galaxies: statistics, infrared: galaxies, Astrophysics},
   year          = 2002,
   month         = sep,
@@ -38,7 +38,7 @@
 @article{Wang2022,
   author        = {{Wang}, Chun and {Huang}, Yang and {Yuan}, Haibo and {Zhang}, Huawei and {Xiang}, Maosheng and {Liu}, Xiaowei},
   title         = {{The Value-added Catalog for LAMOST DR8 Low-resolution Spectra}},
-  journal       = {\apjs},
+  journal       = {Astrophysical Journal, Supplement},
   keywords      = {Fundamental parameters of stars, 555, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies},
   year          = 2022,
   month         = apr,
@@ -138,7 +138,7 @@
 @article{Schlafly2019,
   author        = {{Schlafly}, Edward F. and {Meisner}, Aaron M. and {Green}, Gregory M.},
   title         = {{The unWISE Catalog: Two Billion Infrared Sources from Five Years of WISE Imaging}},
-  journal       = {\apjs},
+  journal       = {Astrophysical Journals},
   keywords      = {catalogs, infrared: general, surveys, techniques: photometric, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
   year          = 2019,
   month         = feb,
@@ -156,7 +156,7 @@
 @article{Skrutskie2006,
   author   = {{Skrutskie}, M.~F. and {Cutri}, R.~M. and {Stiening}, R. and {Weinberg}, M.~D. and {Schneider}, S. and {Carpenter}, J.~M. and {Beichman}, C. and {Capps}, R. and {Chester}, T. and {Elias}, J. and {Huchra}, J. and {Liebert}, J. and {Lonsdale}, C. and {Monet}, D.~G. and {Price}, S. and {Seitzer}, P. and {Jarrett}, T. and {Kirkpatrick}, J.~D. and {Gizis}, J.~E. and {Howard}, E. and {Evans}, T. and {Fowler}, J. and {Fullmer}, L. and {Hurt}, R. and {Light}, R. and {Kopan}, E.~L. and {Marsh}, K.~A. and {McCallon}, H.~L. and {Tam}, R. and {Van Dyk}, S. and {Wheelock}, S.},
   title    = {{The Two Micron All Sky Survey (2MASS)}},
-  journal  = {\aj},
+  journal  = {Astronomical Journal},
   keywords = {Catalogs, Infrared: General, Surveys},
   year     = 2006,
   month    = feb,
@@ -771,7 +771,7 @@
 @article{Rezaei2022,
   author        = {{Rezaei Kh.}, Sara and {Kainulainen}, Jouni},
   title         = {{Three-dimensional Shape Explains Star Formation Mystery of California and Orion A}},
-  journal       = {\apjl},
+  journal       = {Astrophysical Journal, Letters},
   keywords      = {Interstellar medium, Interstellar dust, Molecular clouds, Star forming regions, Star formation, 847, 836, 1072, 1565, 1569, Astrophysics - Astrophysics of Galaxies},
   year          = 2022,
[…]
