### 2024-05-25 #6593 @dfm
@Edenhofer — I've opened a small PR with some minor edits to the manuscript, please take a look and merge or let me know what you think. Please also incorporate the DOIs identified by @Abinashbunty when you have a chance.

Once you've done that:
