+++ paper/paper.bib
@@ -170,7 +170,7 @@
 @article{Wright2010,
   author        = {{Wright}, Edward L. and {Eisenhardt}, Peter R.~M. and {Mainzer}, Amy K. and {Ressler}, Michael E. and {Cutri}, Roc M. and {Jarrett}, Thomas and {Kirkpatrick}, J. Davy and {Padgett}, Deborah and {McMillan}, Robert S. and {Skrutskie}, Michael and {Stanford}, S.~A. and {Cohen}, Martin and {Walker}, Russell G. and {Mather}, John C. and {Leisawitz}, David and {Gautier}, Thomas N., III and {McLean}, Ian and {Benford}, Dominic and {Lonsdale}, Carol J. and {Blain}, Andrew and {Mendez}, Bryan and {Irace}, William R. and {Duval}, Valerie and {Liu}, Fengchuan and {Royer}, Don and {Heinrichsen}, Ingolf and {Howard}, Joan and {Shannon}, Mark and {Kendall}, Martha and {Walsh}, Amy L. and {Larsen}, Mark and {Cardon}, Joel G. and {Schick}, Scott and {Schwalm}, Mark and {Abid}, Mohamed and {Fabinsky}, Beth and {Naes}, Larry and {Tsai}, Chao-Wei},
   title         = {{The Wide-field Infrared Survey Explorer (WISE): Mission Description and Initial On-orbit Performance}},
-  journal       = {\aj},
+  journal       = {Astronomical Journal},
   keywords      = {infrared: general, space vehicles, surveys, Astrophysics - Instrumentation and Methods for Astrophysics},
   year          = 2010,
   month         = dec,
@@ -538,7 +538,7 @@
 @article{Speagle2020,
   author        = {{Speagle}, Joshua S.},
   title         = {{DYNESTY: a dynamic nested sampling package for estimating Bayesian posteriors and evidences}},
-  journal       = {\mnras},
+  journal       = {Monthly Notices of the RAS},
   keywords      = {methods: data analysis, methods: statistical, Astrophysics - Instrumentation and Methods for Astrophysics, Statistics - Computation},
   year          = 2020,
   month         = apr,
@@ -851,7 +851,7 @@
 @article{Green2019,
   author        = {{Green}, Gregory M. and {Schlafly}, Edward and {Zucker}, Catherine and {Speagle}, Joshua S. and {Finkbeiner}, Douglas},
   title         = {{A 3D Dust Map Based on Gaia, Pan-STARRS 1, and 2MASS}},
-  journal       = {\apj},
+  journal       = {Astrophysical Journal},
   keywords      = {Interstellar reddening, Interstellar dust extinction, Galaxy structure, Galaxy stellar content, Interstellar dust, 853, 837, 622, 621, 836, Astrophysics - Astrophysics of Galaxies},
   year          = 2019,
   month         = dec,
@@ -1009,7 +1009,7 @@
 @article{Kuhn2021,
   author        = {{Kuhn}, Michael A. and {de Souza}, Rafael S. and {Krone-Martins}, Alberto and {Castro-Ginard}, Alfred and {Ishida}, Emille E.~O. and {Povich}, Matthew S. and {Hillenbrand}, Lynne A. and {COIN Collaboration}},
   title         = {{SPICY: The Spitzer/IRAC Candidate YSO Catalog for the Inner Galactic Midplane}},
-  journal       = {\apjs},
+  journal       = {Astrophysical Journal, Supplement},
[…]
