### 2024-05-27 e265642
host_callback: replace host_callback with debug.callback
