### 2024-05-27 4e99866
host_callback: remove commented lines
