+++ src/re/conjugate_gradient.py
@@ -232,8 +232,6 @@ def _static_cg(
     _raise_nonposdef=False,  # TODO
     **kwargs
 ) -> CGResults:
-    # from jax.experimental.host_callback import call
-    # from jax.experimental import io_callback
     from jax.debug import callback
     from jax.lax import cond, while_loop
 
@@ -326,7 +324,6 @@ def _static_cg(
                 "resnorm": resnorm,
                 "maxiter": maxiter
             }
-            # call(pp, printable_state, result_shape=None)
             callback(pp, printable_state)
 
         ret = {
@@ -380,7 +377,6 @@ def _static_cg(
             "resnorm": resnorm,
             "maxiter": maxiter
         }
-        # call(pp, printable_state, result_shape=None)
         callback(pp, printable_state)
 
     val = while_loop(continue_condition, cg_single_step, val)
@@ -501,7 +497,6 @@ def _cg_steihaug_subproblem(
     The Hessian itself is not required, and the Hessian does
     not need to be positive semidefinite.
     """
-    # from jax.experimental.host_callback import call
     from jax.debug import callback
     from jax.lax import switch, while_loop
 
@@ -659,7 +654,6 @@ def _cg_steihaug_subproblem(
                 "case": index,
                 "maxiter": maxiter
             }
-            # call(pp, printable_state, result_shape=None)
             callback(pp, printable_state)
 
         return iterp
