### 2024-05-29 4a8d218
newton_cg: stop cg when energy would increase
