### 2024-05-29 63f9933
0_intro: use new ncg option
