### 2024-06-03 2be9edf
paper.bib: Remove unrelated (and unused) references
