+++ paper/paper.bib
@@ -1,210 +1,3 @@
-@article{Green2018,
-  author  = {{Green}, Gregory M.},
-  title   = {{dustmaps: A Python interface for maps of interstellar dust}},
-  journal = {The Journal of Open Source Software},
-  year    = 2018,
-  month   = jun,
-  volume  = {3},
-  number  = {26},
-  pages   = {695},
-  doi     = {10.21105/joss.00695},
-  adsurl  = {https://ui.adsabs.harvard.edu/abs/2018JOSS....3..695G},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-@book{Draine2011,
-  author  = {{Draine}, Bruce T.},
-  title   = {{Physics of the Interstellar and Intergalactic Medium}},
-  year    = 2011,
-  adsurl  = {https://ui.adsabs.harvard.edu/abs/2011piim.book.....D},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-@article{Popescu2002,
-  author        = {{Popescu}, Cristina C. and {Tuffs}, Richard J.},
-  title         = {{The percentage of stellar light re-radiated by dust in late-type Virgo Cluster galaxies}},
-  journal       = {Monthly Notices of the RAS},
-  keywords      = {galaxies: clusters: individual: Virgo Cluster, galaxies: fundamental parameters, galaxies: photometry, galaxies: spiral, galaxies: statistics, infrared: galaxies, Astrophysics},
-  year          = 2002,
-  month         = sep,
-  volume        = {335},
-  number        = {2},
-  pages         = {L41-L44},
-  doi           = {10.1046/j.1365-8711.2002.05881.x},
-  archiveprefix = {arXiv},
-  eprint        = {astro-ph/0208285},
-  primaryclass  = {astro-ph},
-  adsurl        = {https://ui.adsabs.harvard.edu/abs/2002MNRAS.335L..41P},
-  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
-}
-@article{Wang2022,
-  author        = {{Wang}, Chun and {Huang}, Yang and {Yuan}, Haibo and {Zhang}, Huawei and {Xiang}, Maosheng and {Liu}, Xiaowei},
-  title         = {{The Value-added Catalog for LAMOST DR8 Low-resolution Spectra}},
-  journal       = {Astrophysical Journal, Supplement},
-  keywords      = {Fundamental parameters of stars, 555, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies},
-  year          = 2022,
-  month         = apr,
-  volume        = {259},
-  number        = {2},
-  eid           = {51},
-  pages         = {51},
-  doi           = {10.3847/1538-4365/ac4df7},
-  archiveprefix = {arXiv},
-  eprint        = {2201.09442},
-  primaryclass  = {astro-ph.SR},
-  adsurl        = {https://ui.adsabs.harvard.edu/abs/2022ApJS..259...51W},
-  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
-}
-@article{Xiang2022,
[…]
