+++ paper/paper.bib
@@ -1,3 +1,476 @@
+@misc{Arras2019,
+  author        = {{Arras}, Philipp and {Baltac}, Mihai and {Ensslin}, Torsten A. and {Frank}, Philipp and {Hutschenreuter}, Sebastian and {Knollmueller}, Jakob and {Leike}, Reimar H. and {Newrzella}, Max-Niklas and {Platz}, Lukas and {Reinecke}, Martin and {Stadler}, Julia},
+  title         = {{NIFTy5: Numerical Information Field Theory v5}},
+  keywords      = {Software},
+  howpublished  = {Astrophysics Source Code Library, record ascl:1903.008},
+  year          = 2019,
+  month         = 03,
+  eid           = {ascl:1903.008},
+  pages         = {ascl:1903.008},
+  archiveprefix = {ascl},
+  eprint        = {1903.008}
+}
+@article{Arras2019,
+  author        = {{Arras}, Philipp and {Frank}, Philipp and {Leike}, Reimar and {Westermann}, R{\"u}diger and {En{\ss}lin}, Torsten A.},
+  title         = {{Unified radio interferometric calibration and imaging with joint uncertainty quantification}},
+  journal       = {Astronomy \& Astrophysics},
+  keywords      = {techniques: interferometric, methods: statistical, methods: data analysis, instrumentation: interferometers, Astrophysics - Instrumentation and Methods for Astrophysics},
+  year          = 2019,
+  month         = jul,
+  volume        = {627},
+  eid           = {A134},
+  pages         = {A134},
+  doi           = {10.1051/0004-6361/201935555},
+  archiveprefix = {arXiv},
+  eprint        = {1903.11169},
+  primaryclass  = {astro-ph.IM},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2019A&A...627A.134A},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@article{Arras2022,
+  author        = {{Arras}, Philipp and {Frank}, Philipp and {Haim}, Philipp and {Knollm{\"u}ller}, Jakob and {Leike}, Reimar H. and {Reinecke}, Martin and {En{\ss}lin}, Torsten A.},
+  title         = {{Variable structures in M87* from space, time and frequency resolved interferometry}},
+  journal       = {Nature Astronomy},
+  keywords      = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
+  year          = 2022,
+  month         = jan,
+  volume        = {6},
+  pages         = {259-269},
+  doi           = {10.1038/s41550-021-01548-0},
+  archiveprefix = {arXiv},
+  eprint        = {2002.05218},
+  primaryclass  = {astro-ph.IM},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2022NatAs...6..259A},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{Astropy2013,
[…]
