### 2024-06-03 8255fc0
paper.bib: Spell out abbreviated journals
