### 2024-06-03 09e82f9
Fix DOIs
Resolves #22 .
