+++ paper/paper.bib
@@ -207,14 +207,6 @@
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@misc{Edenhofer2023NIFTyRE,
-  author       = {{Edenhofer}, Gordian and {Frank}, Philipp and {Leike}, Reimar H. and {Roth}, Jakob and {Guerdi}, Massin and {Enßlin}, Torsten A.},
-  title        = {{Re-Envisioning Numerical Information Field Theory (NIFTy): A Library for Gaussian Processes and Variational Inference}},
-  keywords     = {Software},
-  year         = 2023,
-  howpublished = {in preparation}
-}
-
 @article{ForemanMackey2013,
   author        = {{Foreman-Mackey}, Daniel and {Hogg}, David W. and {Lang}, Dustin and {Goodman}, Jonathan},
   title         = {{emcee: The MCMC Hammer}},
@@ -645,18 +637,22 @@
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
 @misc{Selig2013,
-  author        = {{Selig}, Marco and {Bell}, Michael R. and {Junklewitz}, Henrik and {Oppermann}, Niels and {Reinecke}, Martin and {Greiner}, Maksim and {Pachajoa}, Carlos and {Ensslin}, Torsten A.},
-  title         = {{NIFTY: A versatile Python library for signal inference}},
-  keywords      = {Software},
-  howpublished  = {Astrophysics Source Code Library, record ascl:1302.013},
+  author        = {{Selig}, M. and {Bell}, M.~R. and {Junklewitz}, H. and {Oppermann}, N. and {Reinecke}, M. and {Greiner}, M. and {Pachajoa}, C. and {En{\ss}lin}, T.~A.},
+  title         = {{NIFTY - Numerical Information Field Theory. A versatile PYTHON library for signal inference}},
+  journal       = {Astronomy \& Astrophysics},
+  keywords      = {methods: data analysis, methods: numerical, methods: statistical, techniques: image processing, Astrophysics - Instrumentation and Methods for Astrophysics, Computer Science - Information Theory, Computer Science - Mathematical Software, Mathematical Physics, Physics - Data Analysis, Statistics and Probability, Statistics - Computation},
   year          = 2013,
-  month         = 02,
-  eid           = {ascl:1302.013},
-  pages         = {ascl:1302.013},
-  archiveprefix = {ascl},
-  eprint        = {1302.013}
+  month         = jun,
+  volume        = {554},
+  eid           = {A26},
+  pages         = {A26},
+  doi           = {10.1051/0004-6361/201321236},
+  archiveprefix = {arXiv},
+  eprint        = {1301.4499},
+  primaryclass  = {astro-ph.IM},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2013A&A...554A..26S},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
 @article{Sivaram2015,
   author        = {{Ambikasaran}, Sivaram and {Foreman-Mackey}, Daniel and {Greengard}, Leslie and {Hogg}, David W. and {O'Neil}, Michael},
   title         = {{Fast Direct Methods for Gaussian Processes}},
