+++ paper/paper.bib
@@ -191,19 +191,20 @@
 }
 
 @article{Edenhofer2023,
-  author        = {{Edenhofer}, Gordian and {Zucker}, Catherine and {Frank}, Philipp and {Saydjari}, Andrew K. and {Speagle}, Joshua S. and {Finkbeiner}, Douglas and {En{\ss}lin}, Torsten},
-  title         = {{A Parsec-Scale Galactic 3D Dust Map out to 1.25 kpc from the Sun}},
-  journal       = {arXiv e-prints},
-  keywords      = {Astrophysics - Astrophysics of Galaxies},
-  year          = 2023,
-  month         = aug,
-  eid           = {arXiv:2308.01295},
-  pages         = {arXiv:2308.01295},
-  doi           = {10.48550/arXiv.2308.01295},
+  author        = {{Edenhofer}, Gordian and {Zucker}, Catherine and {Frank}, Philipp and {Saydjari}, Andrew K. and {Speagle}, Joshua S. and {Finkbeiner}, Douglas and {En{\ss}lin}, Torsten A.},
+  title         = {{A parsec-scale Galactic 3D dust map out to 1.25 kpc from the Sun}},
+  journal       = {Astronomy \& Astrophysics},
+  keywords      = {ISM: clouds, ISM: structure, dust, extinction, Galaxy: structure, methods: statistical, Astrophysics - Astrophysics of Galaxies},
+  year          = 2024,
+  month         = may,
+  volume        = {685},
+  eid           = {A82},
+  pages         = {A82},
+  doi           = {10.1051/0004-6361/202347628},
   archiveprefix = {arXiv},
   eprint        = {2308.01295},
   primaryclass  = {astro-ph.GA},
-  adsurl        = {https://ui.adsabs.harvard.edu/abs/2023arXiv230801295E},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2024A&A...685A..82E},
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
@@ -330,7 +331,7 @@
   author    = {Guardiani, Matteo and Frank, Philipp and Kostić, Andrija and Edenhofer, Gordian and Roth, Jakob and Uhlmann, Berit and Enßlin, Torsten},
   journal   = {PLOS ONE},
   publisher = {Public Library of Science},
-  title     = {Causal, Bayesian, & non-parametric modeling of the SARS-CoV-2 viral load distribution vs. patient’s age},
+  title     = {Causal, Bayesian, & non-parametric modeling of the SARS-CoV-2 viral load distribution vs. patient's age},
   year      = {2022},
   month     = {10},
   volume    = {17},
