### 2024-06-03 NIFTy-PPL/NIFTy#22 @Edenhofer
Thanks for DOIs! I fixed most of them but left some of them as is since the DOIs are references to the arXiv paper and not the journal publication. Unfortunately, some of the ML journals/conferences don't issue DOIs and hence these publications then don't have any even if their corresponding arXiV paper has one.
