+++ paper/paper.bib
@@ -1,15 +1,3 @@
-@misc{Arras2019,
-  author        = {{Arras}, Philipp and {Baltac}, Mihai and {Ensslin}, Torsten A. and {Frank}, Philipp and {Hutschenreuter}, Sebastian and {Knollmueller}, Jakob and {Leike}, Reimar H. and {Newrzella}, Max-Niklas and {Platz}, Lukas and {Reinecke}, Martin and {Stadler}, Julia},
-  title         = {{NIFTy5: Numerical Information Field Theory v5}},
-  keywords      = {Software},
-  howpublished  = {Astrophysics Source Code Library, record ascl:1903.008},
-  year          = 2019,
-  month         = 03,
-  eid           = {ascl:1903.008},
-  pages         = {ascl:1903.008},
-  archiveprefix = {ascl},
-  eprint        = {1903.008}
-}
 @article{Arras2019,
   author        = {{Arras}, Philipp and {Frank}, Philipp and {Leike}, Reimar and {Westermann}, R{\"u}diger and {En{\ss}lin}, Torsten A.},
   title         = {{Unified radio interferometric calibration and imaging with joint uncertainty quantification}},
@@ -26,6 +14,18 @@
   primaryclass  = {astro-ph.IM},
   adsurl        = {https://ui.adsabs.harvard.edu/abs/2019A&A...627A.134A},
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@misc{Arras2019NIFTy,
+  author        = {{Arras}, Philipp and {Baltac}, Mihai and {Ensslin}, Torsten A. and {Frank}, Philipp and {Hutschenreuter}, Sebastian and {Knollmueller}, Jakob and {Leike}, Reimar H. and {Newrzella}, Max-Niklas and {Platz}, Lukas and {Reinecke}, Martin and {Stadler}, Julia},
+  title         = {{NIFTy5: Numerical Information Field Theory v5}},
+  keywords      = {Software},
+  howpublished  = {Astrophysics Source Code Library, record ascl:1903.008},
+  year          = 2019,
+  month         = 03,
+  eid           = {ascl:1903.008},
+  pages         = {ascl:1903.008},
+  archiveprefix = {ascl},
+  eprint        = {1903.008}
 }
 @article{Arras2022,
   author        = {{Arras}, Philipp and {Frank}, Philipp and {Haim}, Philipp and {Knollm{\"u}ller}, Jakob and {Leike}, Reimar H. and {Reinecke}, Martin and {En{\ss}lin}, Torsten A.},
