+++ paper/paper.md
@@ -115,17 +115,17 @@ With \texttt{NIFTy.re} bridging ideas from \texttt{NIFTy} to JAX, we envision ma
 \texttt{NIFTy.re} competes with other GP libraries as well as with probabilistic programming languages and frameworks.
 Compared to GPyTorch [@Hensman2015], GPflow [@Matthews2017], george [@Sivaram2015], or TinyGP [@ForemanMackey2024], \texttt{NIFTy.re} focuses on GP models for structured spaces and does not assume the posterior to be analytically accessible.
 Instead, \texttt{NIFTy.re} tries to approximate the true posterior using VI.
-Compared to classical probabilistic programming languages such as Stan [@Carpenter2017] and frameworks such Pyro [@Bingham2019], NumPyro [@Phan2019], pyMC3 [@Salvatier2016], emcee [@ForemanMackey2013], dynesty [@Speagle2020; @Koposov2023], or BlackJAX [@blackjax2020], \texttt{NIFTy.re} focuses on inference in extremely high-dimensional spaces.
+Compared to classical probabilistic programming languages such as Stan [@Carpenter2017] and frameworks such as Pyro [@Bingham2019], NumPyro [@Phan2019], pyMC3 [@Salvatier2016], emcee [@ForemanMackey2013], dynesty [@Speagle2020; @Koposov2023], or BlackJAX [@blackjax2020], \texttt{NIFTy.re} focuses on inference in extremely high-dimensional spaces.
 \texttt{NIFTy.re} exploits the structure of probabilistic models in its VI techniques [@Frank2021].
 With \texttt{NIFTy.re}, the GP models and the VI machinery are now fully accessible in the JAX ecosystem and \texttt{NIFTy.re} components interact seamlessly with other JAX packages such as BlackJAX and JAXopt/Optax [@Blondel2021; @Deepmind2020Optax].
 
 # Core Components
 
 \texttt{NIFTy.re} brings tried and tested structured GP models and VI algorithms to JAX.
-GP models are highly useful for imaging problems, and VI algorithms are essential to probing high-dimensional posteriors, which are often encountered in imaging problems.
+GP models are highly useful for imaging problems, and VI algorithms are essential to probe high-dimensional posteriors, which are often encountered in imaging problems.
 \texttt{NIFTy.re} infers the parameters of interest from noisy data via a stochastic mapping that goes in the opposite direction, from the parameters of interest to the data.
 
-\texttt{NIFTy} and \texttt{NIFTy.re} build up hierarchical models for the posterior.
+\texttt{NIFTy} and \texttt{NIFTy.re} build up hierarchical models for the posterior inference.
 The log-posterior function reads $\ln{p(\theta|d)} := \ell(d, f(\theta)) + \ln{p}(\theta) + \mathrm{const}$ with log-likelihood $\ell$, forward model $f$ mapping the parameters of interest $\theta$ to the data space, and log-prior $\ln{p(\theta)}$.
 The goal of the inference is to draw samples from the posterior $p(\theta|d)$.
 
@@ -221,9 +221,9 @@ At the core of both MGVI and geoVI lies an alternating procedure in which one sw
 MGVI and geoVI define the variational posterior via samples, specifically, via samples drawn around an expansion point.
 The samples in MGVI and geoVI exploit model-intrinsic knowledge of the posterior's approximate shape, encoded in the Fisher information metric and the prior curvature [@Frank2021].
 
-\texttt{NIFTy.re} implements both MGVI and geoVI and allows for much finer control over the way samples are drawn and updated compared to \texttt{NIFTy}.
[…]
