+++ paper/paper.bib
@@ -15,17 +15,15 @@
   adsurl        = {https://ui.adsabs.harvard.edu/abs/2019A&A...627A.134A},
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
-@misc{Arras2019NIFTy,
-  author        = {{Arras}, Philipp and {Baltac}, Mihai and {Ensslin}, Torsten A. and {Frank}, Philipp and {Hutschenreuter}, Sebastian and {Knollmueller}, Jakob and {Leike}, Reimar H. and {Newrzella}, Max-Niklas and {Platz}, Lukas and {Reinecke}, Martin and {Stadler}, Julia},
-  title         = {{NIFTy5: Numerical Information Field Theory v5}},
-  keywords      = {Software},
-  howpublished  = {Astrophysics Source Code Library, record ascl:1903.008},
-  year          = 2019,
-  month         = 03,
-  eid           = {ascl:1903.008},
-  pages         = {ascl:1903.008},
-  archiveprefix = {ascl},
-  eprint        = {1903.008}
+@software{Arras2019NIFTy,
+  author       = {{Arras}, Philipp and {Baltac}, Mihai and {Ensslin}, Torsten A. and {Frank}, Philipp and {Hutschenreuter}, Sebastian and {Knollmueller}, Jakob and {Leike}, Reimar and {Newrzella}, Max-Niklas and {Platz}, Lukas and {Reinecke}, Martin and {Stadler}, Julia},
+  title        = {{NIFTy5: Numerical Information Field Theory v5}},
+  howpublished = {Astrophysics Source Code Library, record ascl:1903.008},
+  year         = 2019,
+  month        = mar,
+  eid          = {ascl:1903.008},
+  adsurl       = {https://ui.adsabs.harvard.edu/abs/2019ascl.soft03008A},
+  adsnote      = {Provided by the SAO/NASA Astrophysics Data System}
 }
 @article{Arras2022,
   author        = {{Arras}, Philipp and {Frank}, Philipp and {Haim}, Philipp and {Knollm{\"u}ller}, Jakob and {Leike}, Reimar H. and {Reinecke}, Martin and {En{\ss}lin}, Torsten A.},
