### 2024-06-03 3e1c356
paper.tex: Sync
