+++ paper/paper.tex
@@ -381,12 +381,12 @@ data to the uncertainties in the final image. To do this, we require an
 inference machinery that not only handles extremely high-dimensional
 spaces, but one that does so in a statistically rigorous way.
 
-\texttt{NIFTy} is a Bayesian imaging library (Arras et al., 2019; Selig
-et al., 2013; Steininger et al., 2019). It is designed to infer the
-million- to billion-dimensional posterior distribution in the image
-space from noisy input data. At the core of \texttt{NIFTy} lies a set of
-powerful Gaussian Process (GP) models and accurate Variational Inference
-(VI) algorithms.
+\texttt{NIFTy} is a Bayesian imaging library (Arras, Baltac, et al.,
+2019; Selig et al., 2013; Steininger et al., 2019). It is designed to
+infer the million- to billion-dimensional posterior distribution in the
+image space from noisy input data. At the core of \texttt{NIFTy} lies a
+set of powerful Gaussian Process (GP) models and accurate Variational
+Inference (VI) algorithms.
 
 \texttt{NIFTy.re} is a rewrite of \texttt{NIFTy} in JAX (Bradbury et
 al., 2018) with all relevant previous GP models, new, more flexible GP
@@ -404,20 +404,20 @@ significant gains in maintainability of \texttt{NIFTy.re} compared to
 
 We expect \texttt{NIFTy.re} to be highly useful for many imaging
 applications and envision many applications within and outside of
-astrophysics (Arras et al., 2022, 2019; Eberle et al., 2023, 2022; Frank
-et al., 2017; S. Hutschenreuter et al., 2022; Sebastian Hutschenreuter
-et al., 2023; Leike et al., 2020; Leike \& Enßlin, 2019; Mertsch \&
-Phan, 2023; J. Roth et al., 2023; Jakob Roth et al., 2023; Scheel-Platz
-et al., 2023; Tsouros et al., 2024; Welling et al., 2021; Westerkamp et
-al., 2023). \texttt{NIFTy.re} has already been successfully used in two
-galactic tomography publications (Edenhofer et al., 2023; Leike et al.,
-2022). A very early version of \texttt{NIFTy.re} enabled a
-100-billion-dimensional reconstruction using a maximum posterior
-inference. In a newer publication, \texttt{NIFTy.re} was used to infer a
-500-million-dimensional posterior distribution using VI (Knollmüller \&
-Enßlin, 2019). The latter publication extensively used
-\texttt{NIFTy.re}'s GPU support to reduce the runtime by two orders of
-magnitude compared to the CPU. With \texttt{NIFTy.re} bridging ideas
+astrophysics (Arras et al., 2022; Arras, Frank, et al., 2019; Eberle et
+al., 2023, 2022; Frank et al., 2017; S. Hutschenreuter et al., 2022;
+Sebastian Hutschenreuter et al., 2023; Leike et al., 2020; Leike \&
+Enßlin, 2019; Mertsch \& Phan, 2023; J. Roth et al., 2023; Jakob Roth et
+al., 2023; Scheel-Platz et al., 2023; Tsouros et al., 2024; Welling et
+al., 2021; Westerkamp et al., 2023). \texttt{NIFTy.re} has already been
+successfully used in two galactic tomography publications (Edenhofer et
+al., 2024; Leike et al., 2022). A very early version of
+\texttt{NIFTy.re} enabled a 100-billion-dimensional reconstruction using
+a maximum posterior inference. In a newer publication, \texttt{NIFTy.re}
+was used to infer a 500-million-dimensional posterior distribution using
[…]
