### 2024-06-03 c40172c
ChangeLog.md: Announce HT convention can be configured
