### 2024-06-03 #6593 @Edenhofer
Thanks for your all of your feedback! I've updated the references accordingly. Note, I left out the DOI if the journal paper doesn't have one even if there is a arXiv paper with a DOI. I left out the DOI as the cited paper technically has no unique identifier and the arXiv paper might be slightly different and/or outdated. I'm happy to add them in if the consensus is to include DOIs for arXiv papers in this case.

I tagged the most recent release with [v8.5.1](https://github.com/NIFTy-PPL/NIFTy/tree/v8.5.1) and pushed a release to zenodo with a DOI of [10.5281/zenodo.11441976](https://doi.org/10.5281/zenodo.11441976).
