### 2024-06-03 e91598c
optimize_kl: Special case error message for first resume
