### 2024-06-04 f6c119f
re.evi.nonlinearly_update_residual: swap jit and likelihood insertion to enable non-static argument handling
