### 2024-06-04 fecb754
Revert "Merge branch '408-host_callback_depricated' into 'NIFTy_8'"
This reverts commit 47b1fc3c33afbc1e40523a30fa5ffdc7b278d6ce, reversing
changes made to c40172cbbf768954e352f034d54a8b3a3baea562.
