### 2024-06-04 c96db3e
re.model: Introduce VModel
