+++ src/re/model.py
@@ -10,17 +10,21 @@ from warnings import warn
 
 from jax import eval_shape
 from jax import numpy as jnp
-from jax import random
+from jax import random, vmap
 from jax.tree_util import (
-    register_pytree_node, tree_leaves, tree_map, tree_structure, tree_unflatten
+    register_pytree_node,
+    tree_leaves,
+    tree_map,
+    tree_structure,
+    tree_unflatten,
 )
 
 from .misc import wrap
 from .tree_math import PyTreeString, ShapeWithDtype, random_like
 
 
-class Initializer():
-    domain = ShapeWithDtype((2, ), jnp.uint32)
+class Initializer:
+    domain = ShapeWithDtype((2,), jnp.uint32)
 
     def __new__(cls, call_or_struct):
         if isinstance(call_or_struct, Initializer):
@@ -34,9 +38,7 @@ class Initializer():
         if not self.stupid:
             struct = tree_structure(self._call_or_struct)
             # Cast the subkeys to the structure of `primals`
-            subkeys = tree_unflatten(
-                struct, random.split(key, struct.num_leaves)
-            )
+            subkeys = tree_unflatten(struct, random.split(key, struct.num_leaves))
 
             def draw(init, key):
                 return init(key, *args, **kwargs)
@@ -90,6 +92,7 @@ class ModelMeta(abc.ABCMeta):
     For any dataclasses.Field property with a metadata-entry named "static",
     we will either hide or expose the property to JAX depending on the value.
     """
+
     def __new__(mcs, name, bases, dict_, /, **kwargs):
         cls = super().__new__(mcs, name, bases, dict_, **kwargs)
         cls = dataclass(init=False, repr=False, eq=False)(cls)
@@ -121,7 +124,7 @@ class ModelMeta(abc.ABCMeta):
         return cls
 
 
-class NoValue():
+class NoValue:
     pass
 
 
@@ -159,9 +162,7 @@ class LazyModel(metaclass=ModelMeta):
             )
             warn(msg)
             return Initializer(
-                tree_map(
-                    lambda p: partial(random_like, primals=p), self.domain
-                )
+                tree_map(lambda p: partial(random_like, primals=p), self.domain)
             )
         return self._init
 
@@ -175,6 +176,7 @@ class Model(LazyModel):
     metaprogramming. By default all properties are hidden from JAX except those
     marked via `dataclasses.field(metadata=dict(static=False))` as non-static.
     """
+
     def __init__(
         self,
         call: Optional[Callable] = None,
@@ -280,6 +282,55 @@ class WrappedCall(Model):
         if name is not None:
             call = wrap(call, name=name)
             domain = {name: domain}
-        super().__init__(
-            call, domain=domain, target=target, white_init=white_init
-        )
+        super().__init__(call, domain=domain, target=target, white_init=white_init)
+
+
+class VModel(LazyModel):
+    model: LazyModel = field(metadata=dict(static=False))
+    in_axes: Any
+    out_axes: Any
+    axis_size: int
+
+    def __init__(self, model, axis_size, in_axes=0, out_axes=0):
+        if not isinstance(model, LazyModel):
+            raise ValueError(f"Model {model} of invalid type")
+        self.model = model
+
+        if not isinstance(axis_size, int) and axis_size <= 0:
+            raise ValueError(f"Invalid axis size: {axis_size}")
+
+        struct = tree_structure(model.domain)
+        if isinstance(in_axes, int):
+            in_axes = tree_unflatten(struct, (in_axes,) * struct.num_leaves)
+        else:
+            ax_struct = tree_structure(in_axes)
+            if ax_struct != struct:
+                msg = f"Model domain structure {struct} does not match axis structure {ax_struct}"
+                raise ValueError(msg)
+
[…]
