### 2024-06-04 3519d34
re.evi: Also freeze likelihood for trafo_at_p
