+++ src/re/__init__.py
@@ -24,7 +24,7 @@ from .likelihood_impl import (
 from .logger import logger
 from .minisanity import minisanity, reduced_residual_stats
 from .misc import hvp, interpolate, wrap, wrap_left
-from .model import Initializer, Model, WrappedCall
+from .model import Initializer, Model, VModel, WrappedCall
 from .num import *
 from .optimize import minimize, newton_cg, trust_ncg
 from .optimize_kl import OptimizeVI, OptimizeVIState, optimize_kl
