### 2024-06-04 9777a8d
re.vmodel: Allow for key shortcut of dict only models
