### 2024-06-04 10b1bd3
re.vmodel: Test in_axes by key
