### 2024-06-05 b16ac64
test_vmodel: Simplify
