### 2024-06-05 8df09a8
test_vmodel: Fix WrappedCall usage
