+++ src/re/evi.py
@@ -43,16 +43,15 @@ def _hcb_maybe_raise(condition_exception):
 
 
 def _cond_raise(condition, exception):
-    from jax.experimental.host_callback import call
+    from jax.debug import callback
 
     # Register as few host-callbacks as possible by implicitly hashing the
     # exception type and the strings within
-    call(
+    callback(
         _hcb_maybe_raise, (
             condition,
             Partial(exception.__class__, *hide_strings(exception.args))
-        ),
-        result_shape=None
+        )
     )
 
 
