### 2024-06-06 d5162b3
re.special: Add MF model to special models
