### 2024-06-06 7755c6e
re.vmodel: Fix dict handling of VModel
