### 2024-06-06 a995e97
compare to prev impl
