### 2024-06-06 45cbf1c
VModel: More flexible application for "incompatible" domains
