### 2024-06-07 NIFTy-PPL/NIFTy#22 @Edenhofer
@Abinashbunty Are you ok with not adding DOIs to arXiv if the work is published in a journal and technically doesn't have a DOI?
