### 2024-06-10 6a4bec2
misc.py: Expose conditional_raise and use in CG
Co-authored-by: Lukas Scheel-Platz <lplatz@mpa-garching.mpg.de>
