### 2024-06-10 469a198
cg: Fix i==0 case with negative curvature
