### 2024-06-10 ff06941
_static_cg: Implement _raise_nonposdef
