### 2024-06-10 f4d8e78
Implement a static version of Newton CG (NCG)
Co-authored-by: Gordian Edenhofer <gordian.edenhofer@gmail.com>
