+++ test/test_re/test_ncg.py
@@ -6,8 +6,10 @@ import pytest
 
 pytest.importorskip("jax")
 
+from functools import partial
+
 import jax.numpy as jnp
-from jax import random, value_and_grad
+from jax import jit, random, value_and_grad
 from numpy.testing import assert_allclose
 
 import nifty8.re as jft
@@ -48,13 +50,15 @@ def eggholder(np):
     return func
 
 
-def test_ncg_for_pytree():
+@pmp("ncg", (jft.newton_cg, jft.static_newton_cg))
+def test_ncg_for_pytree(ncg):
     pos = jft.Vector(
         [
             jnp.array(0., dtype=jnp.float32),
-            (jnp.array(3., dtype=jnp.float32), ), {
+            (jnp.array(3., dtype=jnp.float32), ),
+            {
                 "a": jnp.array(5., dtype=jnp.float32)
-            }
+            },
         ]
     )
     getters = (lambda x: x[0], lambda x: x[1][0], lambda x: x[2]["a"])
@@ -74,7 +78,7 @@ def test_ncg_for_pytree():
         m.append({"a": tan[2]["a"] * met[2]})
         return jft.Vector(m)
 
-    res = jft.newton_cg(
+    res = ncg(
         fun_and_grad=value_and_grad(model),
         x0=pos,
         hessp=metric,
@@ -86,7 +90,8 @@ def test_ncg_for_pytree():
 
 
 @pmp("seed", (3637, 12, 42))
-def test_ncg(seed):
+@pmp("ncg", (jft.newton_cg, jft.static_newton_cg))
+def test_ncg(seed, ncg):
     key = random.PRNGKey(seed)
     x = random.normal(key, shape=(3, ))
     diag = jnp.array([1., 2., 3.])
@@ -95,7 +100,7 @@ def test_ncg(seed):
         jnp.sum(y**2 / diag) / 2 - jnp.dot(x, y), y / diag - x
     )
 
-    res = jft.newton_cg(
+    res = ncg(
         fun_and_grad=val_and_grad,
         x0=x,
         hessp=met,
@@ -104,6 +109,25 @@ def test_ncg(seed):
         name='N'
     )
     assert_allclose(res, diag * x, rtol=1e-4, atol=1e-4)
+
+
+@pmp("seed", (3637, 12, 42))
+def test_static_ncg_jittability(seed):
+    key = random.PRNGKey(seed)
+    x = random.normal(key, shape=(3, ))
+    diag = jnp.array([1., 2., 3.])
+    met = lambda y, t: t / diag
+    fun = lambda y: jnp.sum(y**2 / diag) / 2 - jnp.dot(x, y)
+    grad = lambda y: y / diag - x
+    fun_and_grad = lambda y: (fun(y), grad(y))
+
+    for kwargs in [{'fun': fun, 'jac': grad}, {'fun_and_grad': fun_and_grad}]:
+        ncg = jit(
+            partial(jft.static_newton_cg, hessp=met, **kwargs),
+            static_argnames=('name', 'absdelta')
+        )
+        res = ncg(x0=x, maxiter=20, absdelta=1e-6, name='N')
+        assert_allclose(res, diag * x, rtol=1e-4, atol=1e-4)
 
 
 @pmp("seed", (3637, 12, 42))
@@ -171,7 +195,7 @@ def test_cg_steihaug_vs_cg_consistency(seed, size):
     # Note, the solution to the subproblem with infinite trust radius is the CG
     # but with the opposite sign
     for i in range(4):
-        print(f"Iteratoin {i:02d}", file=sys.stderr)
+        print(f"Iteration {i:02d}", file=sys.stderr)
[…]
