### 2024-06-10 3765fbb
test_static_ncg_jittability: Relax static req.
