+++ src/re/correlated_field.py
@@ -1,4 +1,5 @@
 # SPDX-License-Identifier: GPL-2.0+ OR BSD-2-Clause
+# Authors: Gordian Edenhofer, Philipp Frank
 
 import operator
 from collections import namedtuple
@@ -30,6 +31,7 @@ def hartley(p, axes=None):
 
 def get_sht(nside, axis, lmax, mmax, nthreads):
     from jaxbind.contrib.jaxducc0 import get_healpix_sht
+
     jsht = get_healpix_sht(nside, lmax, mmax, 0, nthreads)
 
     def f(inp):
@@ -37,8 +39,10 @@ def get_sht(nside, axis, lmax, mmax, nthreads):
         inp = jnp.moveaxis(inp, axis, -1)
         shp = inp.shape
         inp = inp.reshape((-1, shp[-1]))
+
         def trafo(x):
-            return np.sqrt(4*np.pi) * jsht(x[jnp.newaxis])[0][0]
+            return np.sqrt(4 * np.pi) * jsht(x[jnp.newaxis])[0][0]
+
         res = vmap(trafo)(inp).reshape(shp[:-1] + (-1,))
         return jnp.moveaxis(res, -1, axis)
 
@@ -91,7 +95,7 @@ def get_spherical_mode_distributor(nside: int, lmax=None, mmax=None):
         Multiplicity for each unique mode length.
     """
     if lmax is None:
-        lmax=2*nside
+        lmax = 2 * nside
     lmax = int(lmax)
     if lmax < 0:
         raise ValueError("lmax must be >=0.")
@@ -100,17 +104,17 @@ def get_spherical_mode_distributor(nside: int, lmax=None, mmax=None):
     mmax = int(mmax)
     if mmax < 0 or mmax > lmax:
         raise ValueError("mmax must be >=0 and <=lmax.")
-    size = (lmax+1)**2 - (lmax-mmax)*(lmax-mmax+1)
+    size = (lmax + 1) ** 2 - (lmax - mmax) * (lmax - mmax + 1)
 
     ldist = np.empty((size,), dtype=np.float64)
-    ldist[0:lmax+1] = np.arange(lmax+1, dtype=np.float64)
-    tmp = np.empty((2*lmax+2), dtype=np.float64)
-    tmp[0::2] = np.arange(lmax+1)
-    tmp[1::2] = np.arange(lmax+1)
-    idx = lmax+1
-    for m in range(1, mmax+1):
-        ldist[idx:idx+2*(lmax+1-m)] = tmp[2*m:]
-        idx += 2*(lmax+1-m)
+    ldist[0 : lmax + 1] = np.arange(lmax + 1, dtype=np.float64)
+    tmp = np.empty((2 * lmax + 2), dtype=np.float64)
+    tmp[0::2] = np.arange(lmax + 1)
+    tmp[1::2] = np.arange(lmax + 1)
+    idx = lmax + 1
+    for m in range(1, mmax + 1):
+        ldist[idx : idx + 2 * (lmax + 1 - m)] = tmp[2 * m :]
+        idx += 2 * (lmax + 1 - m)
 
     return _unique_mode_distributor(ldist), (lmax, mmax, size)
 
@@ -138,10 +142,10 @@ def get_fourier_mode_distributor(
     mode_multiplicity : jnp.ndarray
         Multiplicity for each unique Fourier mode length.
     """
-    shape = (shape, ) if isinstance(shape, int) else tuple(shape)
+    shape = (shape,) if isinstance(shape, int) else tuple(shape)
 
     # Compute length of modes
-    mspc_distances = 1. / (jnp.array(shape) * jnp.array(distances))
+    mspc_distances = 1.0 / (jnp.array(shape) * jnp.array(distances))
[…]
