### 2024-06-11 8d457a3
re.correlated_field: Add tests for CF on the sphere
