### 2024-06-13 NIFTy-PPL/NIFTy#22 @Edenhofer
@Abinashbunty Are you happy with this logic of adding DOIs or do you see a better avenue here? In case of the former, could you please confirm in the NIFTy JOSS review that you are okay with it.
