### 2024-06-13 NIFTy-PPL/NIFTy#22 @Abinashbunty
Hi @Edenhofer It is fine by me. 👍🏻 I know the references should be some journal at least and also verified that they are indeed published but don't have a DOI. I just wanted to make sure the bot doesn't complain.

Nevertheless, I approve. 😄
