### 2024-06-14 cd53903
Stub for multi-grid testing code
For now this is mostly just my playground for debugging.
