### 2024-06-14 0d34649
HEALPixGridAtLevel.neighborhood: Align shapes
