### 2024-06-14 ba4e31e
Implement MG ICR refinement matrix retrieval
