### 2024-06-14 5012cc3
HEALPixGridAtLevel.neighborhood: Fix valid for all neighbors
