### 2024-06-14 e9ac298
MG ICR: Couple everything in zeroth layer
