### 2024-06-14 bc47957
MG ICR: Implement (maybe correct) refinement
