+++ test/test_re/test_multi_grid.py
@@ -7,7 +7,7 @@ pytest.importorskip("jax")
 import jax
 import jax.numpy as jnp
 import numpy as np
-from jax import random, value_and_grad
+from jax import random
 from numpy.testing import assert_allclose
 
 import nifty8.re as jft
@@ -34,9 +34,21 @@ if __name__ == "__main__":
     mapped_kernel = jax.vmap(mapped_kernel, in_axes=(None, -1), out_axes=-1)
     mapped_kernel = jax.vmap(mapped_kernel, in_axes=(-1, None), out_axes=-1)
 
+    grid_at_lvl = grid.at(0)
+    # TODO: Children of multiple pixels could be refined jointly requiring a
+    # non-unit stride
+    pixel_indices = np.mgrid[tuple(slice(0, sz) for sz in grid_at_lvl.shape)]
+    g_shape = pixel_indices.shape[1:]
+    assert pixel_indices.shape[0] == grid_at_lvl.ndim
+    pixel_indices = pixel_indices.reshape(grid_at_lvl.ndim, -1)
+    gc = grid_at_lvl.index2coord(pixel_indices)
+    gc = gc.reshape(gc.shape[0], np.prod(g_shape))
+    cov = mapped_kernel(gc, gc)
+    assert cov.shape == (np.prod(g_shape),) * 2
+    cov_sqrt0 = jnp.linalg.cholesky(cov)
+
     for lvl in range(grid.depth):
         grid_at_lvl = grid.at(lvl)
-        assert np.all(grid_at_lvl.splits % 2 == 0)
         # TODO: Children of multiple pixels could be refined jointly requiring a
         # non-unit stride
         pixel_indices = np.mgrid[tuple(slice(0, sz) for sz in grid_at_lvl.shape)]
@@ -44,34 +56,86 @@ if __name__ == "__main__":
         assert pixel_indices.shape[0] == grid_at_lvl.ndim
         pixel_indices = pixel_indices.reshape(grid_at_lvl.ndim, -1)
 
-        ws = window_size if lvl != 0 else g_shape
-
         # Compute neighbors and map to coordinates outside of vmap to support
         # non-batchable (non-JAX) coordinate transformations
-        gc, _ = grid_at_lvl.neighborhood(pixel_indices, ws)
+        gc, _ = grid_at_lvl.neighborhood(pixel_indices, window_size)
         gc = grid_at_lvl.index2coord(gc)
         gf = grid_at_lvl.children(pixel_indices)
         gf = grid.at(lvl + 1).index2coord(gf)
         assert gc.shape[0] == gf.shape[0]
         odim = gc.shape[0]
-        gc = gc.reshape(odim, np.prod(g_shape), np.prod(ws))
+        gc = gc.reshape(odim, np.prod(g_shape), np.prod(window_size))
         gf = gf.reshape(odim, np.prod(g_shape), np.prod(grid_at_lvl.splits))
         coord = jnp.concatenate((gc, gf), axis=-1)
         del gc, gf
         cov = mapped_kernel(coord, coord)
         del coord
-        _cs = np.prod(ws) + np.prod(grid_at_lvl.splits)
+        _cs = np.prod(window_size) + np.prod(grid_at_lvl.splits)
         assert cov.shape == (np.prod(g_shape),) + (_cs,) * 2
 
-        olf, ks = jax.vmap(refinement_matrices, in_axes=(0, None, None))(
+        olf, fks = jax.vmap(refinement_matrices, in_axes=(0, None, None))(
             cov, np.prod(grid_at_lvl.splits), coerce_fine_kernel
         )
-        olf = olf.reshape(g_shape + olf.shape[-2:])
[…]
