### 2024-06-14 7c4b9e1
MG ICR: Plot final result
